Integrate Stream review readability improvements
