+++ i18n/messages/public-review-artists-and-roles.ts
@@ -1,0 +1,530 @@
+export const PUBLIC_REVIEW_ARTISTS_AND_ROLES_MESSAGES = {
+  "publicReview.forArtistsGuide.artwork.heading":
+    "Understand the plan before you approve it",
+  "publicReview.forArtistsGuide.artwork.description":
+    "Before launch, check the artwork, edition size, sale, payments, and who can still make changes. Stream records the exact version you approve.",
+  "publicReview.forArtistsGuide.artwork.identity.title":
+    "What artwork are you publishing?",
+  "publicReview.forArtistsGuide.artwork.identity.description":
+    "Choose the files, scripts, and other parts the artwork needs. Stream records the artist, the artwork, and the tokens created from it.",
+  "publicReview.forArtistsGuide.artwork.editions.title":
+    "Is it unique or an edition?",
+  "publicReview.forArtistsGuide.artwork.editions.description":
+    "Choose whether there is one token or an edition, how many tokens can exist, and how each token is created. If the work uses randomness, Stream saves the result for each token so it can be recreated later.",
+  "publicReview.forArtistsGuide.artwork.sales.title":
+    "How can collectors get it?",
+  "publicReview.forArtistsGuide.artwork.sales.description":
+    "The plan shows whether collectors mint at a fixed price or bid in an auction. It also shows the sale rules and who can take part.",
+  "publicReview.forArtistsGuide.artwork.payments.title":
+    "Where does the money go?",
+  "publicReview.forArtistsGuide.artwork.payments.description":
+    "Check the price, currency, and everyone who receives a share. Stream can state a royalty for later sales, but outside marketplaces may not pay it.",
+  "publicReview.forArtistsGuide.artwork.approval.title":
+    "What are you approving?",
+  "publicReview.forArtistsGuide.artwork.approval.description":
+    "Your wallet approves one exact version of the plan. If important details change, you need to review and approve the new version.",
+  "publicReview.forArtistsGuide.artwork.control.title":
+    "What can change or become permanent?",
+  "publicReview.forArtistsGuide.artwork.control.description":
+    "Before finality, some files, supply, minting, and sale settings can still change. Stream keeps a record of those changes. The finality process is meant to close the remaining ways the artwork itself can change.",
+  "publicReview.forArtistsGuide.journey.heading":
+    "How your artwork moves through Stream",
+  "publicReview.forArtistsGuide.journey.description":
+    "Your artwork moves through six stages. Each stage has different choices, checks, and approvals.",
+  "publicReview.forArtistsGuide.journey.listLabel": "Artwork journey stages",
+  "publicReview.forArtistsGuide.journey.prepare.title": "Prepare",
+  "publicReview.forArtistsGuide.journey.prepare.description":
+    "Add the artwork files and anything they need to work. Choose the edition size, sale method, credits, storage, and payment shares.",
+  "publicReview.forArtistsGuide.journey.review.title": "Review",
+  "publicReview.forArtistsGuide.journey.review.description":
+    "Check the complete plan. Make sure the artwork, sale, payments, and who can still make changes are correct.",
+  "publicReview.forArtistsGuide.journey.approve.title": "Approve",
+  "publicReview.forArtistsGuide.journey.approve.description":
+    "Use your wallet to approve one exact version of the plan. If an important detail changes, review and approve the new version.",
+  "publicReview.forArtistsGuide.journey.select.title": "Community selection",
+  "publicReview.forArtistsGuide.journey.select.description":
+    "The community reviews the work outside the contracts. If it is selected, a separate Stream wallet signs the exact mint or auction terms.",
+  "publicReview.forArtistsGuide.journey.launch.title": "Launch",
+  "publicReview.forArtistsGuide.journey.launch.description":
+    "Stream checks the signed permission. Collectors can then mint or bid using only those sale terms.",
+  "publicReview.forArtistsGuide.journey.finalize.title": "Finalize",
+  "publicReview.forArtistsGuide.journey.finalize.description":
+    "Finality has a waiting period. During this time, the artist and reviewers check the files, supply, and every remaining way the artwork could change. The final step is meant to close those remaining change paths.",
+  "publicReview.forArtistsGuide.journey.important":
+    "Important: Ending minting, freezing settings, recording preservation evidence, and finalizing the artwork are separate steps. Completing one does not complete the others.",
+  "publicReview.forArtistsGuide.approval.heading": "Before you approve",
+  "publicReview.forArtistsGuide.approval.description":
+    "Your wallet approval covers specific details, not every part of Stream. Check what the signature covers and review the wider artwork plan before signing.",
+  "publicReview.forArtistsGuide.approval.listLabel": "Approval checklist",
+  "publicReview.forArtistsGuide.approval.signing.title": "What you are signing",
+  "publicReview.forArtistsGuide.approval.signing.description":
+    "Check the artwork, artist wallet, contract, network, and version.",
+  "publicReview.forArtistsGuide.approval.artworkSupply.title":
+    "Artwork and supply",
+  "publicReview.forArtistsGuide.approval.artworkSupply.description":
+    "Check every required file and the maximum number of tokens. Include every way more tokens can be created.",
+  "publicReview.forArtistsGuide.approval.salePayments.title":
+    "Sale and payments",
+  "publicReview.forArtistsGuide.approval.salePayments.description":
+    "Check the sale type, price, currency, refunds, payment recipients, and shares.",
+  "publicReview.forArtistsGuide.approval.powerFinality.title":
+    "Power and finality",
+  "publicReview.forArtistsGuide.approval.powerFinality.description":
+    "Check who can change, pause, or restart each part. Check what every permanent step locks.",
+  "publicReview.forArtistsGuide.approval.note":
+    "If a signed detail changes, the old approval stays with the old version. Review the new version before signing again.",
+  "publicReview.forArtistsGuide.changes.heading": "What can still change?",
+  "publicReview.forArtistsGuide.changes.description":
+    "Stream makes parts of your artwork permanent at different times. Ending minting, freezing settings, and artwork finality are separate steps.",
+  "publicReview.forArtistsGuide.changes.beforeLaunch.title":
+    "Before minting starts",
+  "publicReview.forArtistsGuide.changes.beforeLaunch.description":
+    "Review the plan. If something is wrong, do not approve it. A corrected version can be prepared for you.",
+  "publicReview.forArtistsGuide.changes.afterLaunch.title":
+    "After minting starts",
+  "publicReview.forArtistsGuide.changes.afterLaunch.description":
+    "Minted tokens, payments, and signed actions stay in the history. Some settings can still change. Important changes may need a new approval from you.",
+  "publicReview.forArtistsGuide.changes.afterFinality.title":
+    "After artwork finality",
+  "publicReview.forArtistsGuide.changes.afterFinality.description":
+    "Finality is designed to lock the artwork records it covers. Before this happens, a waiting period lets the guardian stop a suspicious finality action.",
+  "publicReview.forArtistsGuide.actors.heading":
+    "Who else can affect your artwork?",
+  "publicReview.forArtistsGuide.actors.description":
+    "Your approval matters, but other people and services also have a role. Before you sign, check what each one can do and where its power ends.",
+  "publicReview.forArtistsGuide.actors.community.title": "Community review",
+  "publicReview.forArtistsGuide.actors.community.description":
+    "The community reviews and selects the work outside the contracts. The contracts cannot tell whether this decision was fair or correct.",
+  "publicReview.forArtistsGuide.actors.signer.title": "Stream signing wallet",
+  "publicReview.forArtistsGuide.actors.signer.description":
+    "If the work is selected, a separate Stream wallet signs the exact mint or auction terms. The contracts can check that signature.",
+  "publicReview.forArtistsGuide.actors.operators.title":
+    "People who run Stream",
+  "publicReview.forArtistsGuide.actors.operators.description":
+    "The contracts give different roles different powers. Check who can change, pause, or restart each part of Stream.",
+  "publicReview.forArtistsGuide.actors.guardian.title": "Guardian",
+  "publicReview.forArtistsGuide.actors.guardian.description":
+    "The guardian is a safety brake. It can pause a specific part of Stream or stop a planned change during its waiting period. It cannot replace that change with different artwork or terms.",
+  "publicReview.forArtistsGuide.actors.services.title": "Outside services",
+  "publicReview.forArtistsGuide.actors.services.description":
+    "File storage, websites, services that provide random results, and marketplaces work outside the contracts. Stream can record links and evidence. It cannot keep a service online or force a marketplace to pay royalties.",
+  "publicReview.forArtistsGuide.sales.heading": "How sales and payments work",
+  "publicReview.forArtistsGuide.sales.buy.title": "How collectors buy",
+  "publicReview.forArtistsGuide.sales.buy.description":
+    "Collectors pay the signed fixed price or bid in an auction. The current signed sale paths use ETH. Before launch, check the auction end time, cancellation rules, and refund rules.",
+  "publicReview.forArtistsGuide.sales.money.title": "Where the money goes",
+  "publicReview.forArtistsGuide.sales.money.description":
+    "Check every person or wallet that receives money and their share. After the sale, Stream records how much each one can withdraw. Each person withdraws separately, so one failed payment does not block everyone else.",
+  "publicReview.forArtistsGuide.sales.royalties.title": "Later royalties",
+  "publicReview.forArtistsGuide.sales.royalties.description":
+    "Stream can show royalty information for later marketplace sales. The marketplace decides whether to pay it.",
+  "publicReview.forArtistsGuide.permanence.heading":
+    "Before your artwork becomes final",
+  "publicReview.forArtistsGuide.permanence.description":
+    "Finality is meant to lock the artwork records it covers. Ending minting, freezing settings, and adding preservation records are separate steps.",
+  "publicReview.forArtistsGuide.permanence.files.title":
+    "Check every required part",
+  "publicReview.forArtistsGuide.permanence.files.description":
+    "Make sure every file, script, font, and other required part can still be found and opened.",
+  "publicReview.forArtistsGuide.permanence.fingerprints.title":
+    "Check the recorded fingerprints",
+  "publicReview.forArtistsGuide.permanence.fingerprints.description":
+    "A file fingerprint is a code made from the file. A matching code shows that the file has not changed. It does not keep the file online.",
+  "publicReview.forArtistsGuide.permanence.record.title":
+    "Check the final record",
+  "publicReview.forArtistsGuide.permanence.record.description":
+    "Confirm the token supply, mint history, payments, and any saved randomness results.",
+  "publicReview.forArtistsGuide.permanence.delay.title":
+    "Use the waiting period",
+  "publicReview.forArtistsGuide.permanence.delay.description":
+    "Finality has a delay so you and other reviewers can find mistakes. A guardian can stop the action during this time, but cannot replace it with different artwork or terms.",
+  "publicReview.forArtistsGuide.permanence.reviewStatus.title":
+    "What is still under review",
+  "publicReview.forArtistsGuide.permanence.reviewStatus.description":
+    "Stream is meant to close every approved path that can change the artwork. The current review has not yet proved that every path is closed. Check the technical evidence before relying on finality.",
+  "publicReview.forArtistsGuide.nextStep.heading": "Your next step",
+  "publicReview.forArtistsGuide.nextStep.description":
+    "Review the exact approval package. If it matches the plan you accept, sign it. If anything is missing or unclear, ask for a corrected version.",
+  "publicReview.forArtistsGuide.readPage": "Read {page}",
+  "publicReview.forArtistsGuide.evidence.heading": "Read the contract details",
+  "publicReview.forArtistsGuide.evidence.description":
+    "The short sections below show what is in the reviewed code, what comes from accepted design decisions, and what is still proposed.",
+  "publicReview.forArtistsGuide.evidence.action":
+    "Continue to the contract details",
+  "publicReview.forArtistsDetails.heading": "How to read the details",
+  "publicReview.forArtistsDetails.description":
+    "The labels show what kind of evidence each section uses. A section can use more than one.",
+  "publicReview.forArtistsDetails.basis.code.label": "Reviewed code",
+  "publicReview.forArtistsDetails.basis.code.description":
+    "Includes behavior present in the pinned contract snapshot.",
+  "publicReview.forArtistsDetails.basis.accepted.label": "Accepted design",
+  "publicReview.forArtistsDetails.basis.accepted.description":
+    "Includes decisions approved in an ADR. They may not be complete in the pinned code.",
+  "publicReview.forArtistsDetails.basis.proposed.label": "Still proposed",
+  "publicReview.forArtistsDetails.basis.proposed.description":
+    "Includes ideas that are not approved or safe to rely on yet.",
+  "publicReview.forArtistsDetails.identity.title":
+    "Your collection has a durable identity",
+  "publicReview.forArtistsDetails.identity.intro":
+    "Stream is designed around one shared ERC-721 Core contract. It gives every token a global ID and a serial number inside its collection.",
+  "publicReview.forArtistsDetails.identity.point.artist":
+    "The Core records the collection and its artist wallet.",
+  "publicReview.forArtistsDetails.identity.point.burn":
+    "The collection link stays in the contract's history after a token is burned.",
+  "publicReview.forArtistsDetails.identity.point.modules":
+    "Sale, rendering, and randomness contracts can change without changing this Core identity.",
+  "publicReview.forArtistsDetails.identity.outro":
+    "Stream is not deployed yet. This describes the reviewed design, not a live guarantee.",
+  "publicReview.forArtistsDetails.approval.title":
+    "Approving a specific collection state",
+  "publicReview.forArtistsDetails.approval.intro":
+    "The current artist signature approves one exact collection state. It also binds the network and Core contract.",
+  "publicReview.forArtistsDetails.approval.point.artist": "Artist wallet",
+  "publicReview.forArtistsDetails.approval.point.freeze":
+    "Collection freeze fingerprint",
+  "publicReview.forArtistsDetails.approval.point.purchases":
+    "Maximum collection purchases",
+  "publicReview.forArtistsDetails.approval.point.supply":
+    "Total collection supply",
+  "publicReview.forArtistsDetails.approval.point.delay":
+    "Delay before final supply",
+  "publicReview.forArtistsDetails.approval.outro":
+    "If one value changes, the old signature stays with the old state. This signature does not directly approve every file, sale rule, payment share, or finality action. Check those separately.",
+  "publicReview.forArtistsDetails.scope.title": "The scope of artist approval",
+  "publicReview.forArtistsDetails.scope.intro":
+    "Today's signature is narrow. The wider design is still deciding which other actions need a fresh artist signature.",
+  "publicReview.forArtistsDetails.scope.point.files":
+    "Artwork files and manifests",
+  "publicReview.forArtistsDetails.scope.point.supply": "Final supply",
+  "publicReview.forArtistsDetails.scope.point.sale": "Sale and payment rules",
+  "publicReview.forArtistsDetails.scope.point.randomness":
+    "Randomness provider",
+  "publicReview.forArtistsDetails.scope.point.freeze": "Core freeze",
+  "publicReview.forArtistsDetails.scope.point.finality": "Artwork finality",
+  "publicReview.forArtistsDetails.scope.point.recovery":
+    "Recovery that changes what viewers receive",
+  "publicReview.forArtistsDetails.scope.outro":
+    "These are open design questions, not current guarantees. A guardian may stop a planned action, but should not be able to replace it with different artwork.",
+  "publicReview.forArtistsDetails.statements.title":
+    "Statements made in the artist's name",
+  "publicReview.forArtistsDetails.statements.intro":
+    "Stream separates records by who is allowed to speak: artist, owner, institution, independent reviewer, and other roles.",
+  "publicReview.forArtistsDetails.statements.rules":
+    "The reviewed code limits who can submit protected record types. An ordinary admin cannot automatically submit an artist statement.",
+  "publicReview.forArtistsDetails.statements.point.author":
+    "Who submitted the statement",
+  "publicReview.forArtistsDetails.statements.point.authority":
+    "Which wallet or authority allowed it",
+  "publicReview.forArtistsDetails.statements.point.subject":
+    "Which collection or token it concerns",
+  "publicReview.forArtistsDetails.statements.point.change":
+    "Whether it can be revised, locked, or replaced by a newer record",
+  "publicReview.forArtistsDetails.statements.outro":
+    "A record can prove who submitted a claim. It cannot prove that the claim is true. Rights, archive, and provenance claims still need outside evidence.",
+  "publicReview.forArtistsDetails.files.title":
+    "Artwork files, scripts, and token data",
+  "publicReview.forArtistsDetails.files.intro":
+    "The artwork may depend on more than one image. It can include scripts, fonts, libraries, token data, and outside files.",
+  "publicReview.forArtistsDetails.files.check":
+    "Before approval or finality, ask for one clear file list. It should show:",
+  "publicReview.forArtistsDetails.files.point.parts":
+    "Every required file and script",
+  "publicReview.forArtistsDetails.files.point.order":
+    "The order in which scripts run",
+  "publicReview.forArtistsDetails.files.point.versions":
+    "The version of each dependency",
+  "publicReview.forArtistsDetails.files.point.locations":
+    "Where each file can be found",
+  "publicReview.forArtistsDetails.files.point.hashes":
+    "A fingerprint for the exact bytes",
+  "publicReview.forArtistsDetails.files.point.rebuild":
+    "Instructions for recreating the work",
+  "publicReview.forArtistsDetails.files.outro":
+    "A matching fingerprint shows that a file has not changed. It does not keep the file available.",
+  "publicReview.forArtistsDetails.editions.title": "One-of-ones and editions",
+  "publicReview.forArtistsDetails.editions.intro":
+    "The Core can hold a one-of-one or an edition. Check:",
+  "publicReview.forArtistsDetails.editions.point.maximum":
+    "The maximum number of tokens",
+  "publicReview.forArtistsDetails.editions.point.minted":
+    "How many tokens have ever been minted",
+  "publicReview.forArtistsDetails.editions.point.live":
+    "How many live tokens remain after burns",
+  "publicReview.forArtistsDetails.editions.point.paths":
+    "Every remaining way to mint",
+  "publicReview.forArtistsDetails.editions.point.control":
+    "Who can change supply or mint rules",
+  "publicReview.forArtistsDetails.editions.point.close":
+    "What closes minting for good",
+  "publicReview.forArtistsDetails.editions.outro":
+    "Burning a token removes its ownership record. It does not erase mint history or restore the used mint allowance.",
+  "publicReview.forArtistsDetails.mint.title": "Choosing who can mint",
+  "publicReview.forArtistsDetails.mint.intro":
+    "Artwork identity and mint rules are separate. Mint rules can include times, supply limits, wallet limits, eligibility checks, and an approved signer.",
+  "publicReview.forArtistsDetails.mint.authorization":
+    "In the current signed Drop path, one authorization can mint only one token. An edition needs several authorizations or another approved mint path.",
+  "publicReview.forArtistsDetails.mint.outro":
+    "Before launch, check who can authorize a mint and how used, expired, or cancelled permissions are blocked.",
+  "publicReview.forArtistsDetails.curation.title": "Curation and TDH",
+  "publicReview.forArtistsDetails.curation.intro":
+    "Community curation and TDH happen outside the contracts.",
+  "publicReview.forArtistsDetails.curation.signature":
+    "If a work is selected, a Stream signing wallet signs the exact mint or auction terms. The contract checks that signature.",
+  "publicReview.forArtistsDetails.curation.limit":
+    "The contract can check that the signed terms were not changed. It cannot check whether the community process was fair or the selection was correct.",
+  "publicReview.forArtistsDetails.sales.title":
+    "Fixed-price sales and auctions",
+  "publicReview.forArtistsDetails.sales.intro":
+    "The reviewed sale paths use ETH. Before launch, check:",
+  "publicReview.forArtistsDetails.sales.point.fixed":
+    "Fixed price: token, price, payer, recipient, deadline, payment shares, and cancellation state",
+  "publicReview.forArtistsDetails.sales.point.auction":
+    "Auction: reserve, end time, extension rule, token custody, refunds, cancellation boundary, and no-bid result",
+  "publicReview.forArtistsDetails.sales.point.cancel":
+    "An auction can be cancelled only before its first valid bid and before it ends",
+  "publicReview.forArtistsDetails.sales.outro":
+    "The contracts record proceeds and refunds as credits. Each person withdraws their own credit later. One failed withdrawal does not block the sale.",
+  "publicReview.forArtistsDetails.revenue.title":
+    "Revenue, collaborators, and royalties",
+  "publicReview.forArtistsDetails.revenue.intro":
+    "Primary-sale money can be shared among the artist, collaborators, curator, protocol, and other recipients.",
+  "publicReview.forArtistsDetails.revenue.paths":
+    "The reviewed code has native ETH credit paths for Drops and Auctions. It also has a separate revenue resolver and split-wallet foundation. These are different payment paths.",
+  "publicReview.forArtistsDetails.revenue.point.people":
+    "Every recipient and share",
+  "publicReview.forArtistsDetails.revenue.point.profile":
+    "Which payment profile applies",
+  "publicReview.forArtistsDetails.revenue.point.change":
+    "Who can change the profile and when",
+  "publicReview.forArtistsDetails.revenue.point.rounding":
+    "How rounding, remaining funds, and withdrawals work",
+  "publicReview.forArtistsDetails.revenue.point.royalty":
+    "The royalty receiver and rate",
+  "publicReview.forArtistsDetails.revenue.outro":
+    "Stream can publish ERC-2981 royalty information. Marketplaces choose whether to pay it.",
+  "publicReview.forArtistsDetails.randomness.title": "Randomness",
+  "publicReview.forArtistsDetails.randomness.intro":
+    "If the artwork uses randomness, check the provider and failure plan before minting.",
+  "publicReview.forArtistsDetails.randomness.point.state":
+    "The reviewed code records whether a request is waiting, complete, stale, or received but not fully processed",
+  "publicReview.forArtistsDetails.randomness.point.result":
+    "It stores a fingerprint of the provider result and the seed made from it",
+  "publicReview.forArtistsDetails.randomness.point.retry":
+    "If local processing fails, a retry uses the same saved seed. It does not ask for a new random result",
+  "publicReview.forArtistsDetails.randomness.point.control":
+    "Check who can change the provider, what happens to unfinished requests, and who pays provider fees",
+  "publicReview.forArtistsDetails.freeze.title": "Freezing the work",
+  "publicReview.forArtistsDetails.freeze.intro":
+    "Four steps are easy to confuse:",
+  "publicReview.forArtistsDetails.freeze.point.supply":
+    "Final supply: no more tokens can be minted",
+  "publicReview.forArtistsDetails.freeze.point.core":
+    "Core freeze: locks the collection fields covered by the Core contract",
+  "publicReview.forArtistsDetails.freeze.point.preservation":
+    "Preservation records: save evidence about the files needed to understand or recreate the work",
+  "publicReview.forArtistsDetails.freeze.point.finality":
+    "Artwork finality: a delayed step intended to close the remaining ways the artwork can change",
+  "publicReview.forArtistsDetails.freeze.separate":
+    "These are separate. Completing one does not complete the others.",
+  "publicReview.forArtistsDetails.freeze.check":
+    "Before freeze or finality, check the files, fingerprints, supply, randomness, payments, open mint and sale paths, and who can still make changes.",
+  "publicReview.forArtistsDetails.freeze.limit":
+    "Accepted ADRs describe stronger freeze and finality goals. The current public review has not yet proved that every artwork-changing path is closed. Do not rely on finality until the technical review is complete.",
+  "publicReview.forArtistsDetails.lifetime.title":
+    "Collaborators, delegation, recovery, and estates",
+  "publicReview.forArtistsDetails.lifetime.intro":
+    "The current artist approval is not a complete lifetime authority system. A future system may need:",
+  "publicReview.forArtistsDetails.lifetime.point.collaborators":
+    "Collaborators with narrow roles",
+  "publicReview.forArtistsDetails.lifetime.point.delegation":
+    "Delegated signing powers",
+  "publicReview.forArtistsDetails.lifetime.point.recovery":
+    "Wallet recovery and key changes",
+  "publicReview.forArtistsDetails.lifetime.point.estates":
+    "Estate instructions and successor roles",
+  "publicReview.forArtistsDetails.lifetime.point.disputes":
+    "Rules for disputes and sanctions",
+  "publicReview.forArtistsDetails.lifetime.status":
+    "ADR 0022's registry and helper design is still proposed. It is not approved or implemented.",
+  "publicReview.forArtistsDetails.lifetime.question":
+    "The key question: can someone recover access without gaining power to replace the artwork or rewrite its history?",
+  "publicReview.forArtistsDetails.design.title": "Design position",
+  "publicReview.forArtistsDetails.design.intro":
+    "Artists should never have to sign a hash they cannot understand. Before a wallet asks for a signature, the screen should show:",
+  "publicReview.forArtistsDetails.design.point.work":
+    "The exact artwork and version",
+  "publicReview.forArtistsDetails.design.point.files":
+    "The files and token supply",
+  "publicReview.forArtistsDetails.design.point.sale":
+    "The sale and payment rules",
+  "publicReview.forArtistsDetails.design.point.randomness":
+    "The randomness provider",
+  "publicReview.forArtistsDetails.design.point.power": "Who can change what",
+  "publicReview.forArtistsDetails.design.point.permanent":
+    "What becomes permanent",
+  "publicReview.forArtistsDetails.design.outro":
+    "The goal is simple: every important choice should be visible before it becomes final.",
+  "publicReview.forArtistsDetails.questions.title": "Questions for artists",
+  "publicReview.forArtistsDetails.questions.intro":
+    "These decisions are still worth discussing:",
+  "publicReview.forArtistsDetails.questions.point.signature":
+    "Which actions must need your signature?",
+  "publicReview.forArtistsDetails.questions.point.shared":
+    "Which actions should also need approval from the people who run Stream?",
+  "publicReview.forArtistsDetails.questions.point.collaborators":
+    "What should a collaborator be allowed to do?",
+  "publicReview.forArtistsDetails.questions.point.recovery":
+    "What recovery and estate process would you trust?",
+  "publicReview.forArtistsDetails.questions.point.preservation":
+    "What evidence would you need before freezing the work?",
+  "publicReview.forArtistsDetails.questions.point.payments":
+    "Which sale and payment settings should become permanent?",
+  "publicReview.forArtistsDetails.questions.point.finality":
+    "Would the finality process give you confidence, or add too much work?",
+  "publicReview.rolesGuide.status.heading": "Start with status",
+  "publicReview.rolesGuide.status.description":
+    "Stream is still being prepared. A role may exist in the code without being active in the current rehearsal. Check the label before treating a power as available.",
+  "publicReview.rolesGuide.status.working.title": "Working in rehearsal",
+  "publicReview.rolesGuide.status.working.description":
+    "Connected to the current rehearsal path.",
+  "publicReview.rolesGuide.status.connected.title": "Connected for integration",
+  "publicReview.rolesGuide.status.connected.description":
+    "Connected to selected contracts, but not used by every current path.",
+  "publicReview.rolesGuide.status.source.title": "Built in source",
+  "publicReview.rolesGuide.status.source.description":
+    "The code exists. Final accounts, settings, and release wiring are not proven.",
+  "publicReview.rolesGuide.status.planned.title": "Planned",
+  "publicReview.rolesGuide.status.planned.description":
+    "The design is accepted. Complete code or wiring is still missing.",
+  "publicReview.rolesGuide.status.open.title": "Still open",
+  "publicReview.rolesGuide.status.open.description":
+    "No final design decision has been made.",
+  "publicReview.rolesGuide.current.heading": "Working in the current rehearsal",
+  "publicReview.rolesGuide.current.description":
+    "These are the main powers used by the current signed Drop and auction path.",
+  "publicReview.rolesGuide.current.artist.title": "Artist",
+  "publicReview.rolesGuide.current.artist.description":
+    "Signs one exact collection state. The signature covers specific collection facts, not every later action.",
+  "publicReview.rolesGuide.current.collector.title": "Collector",
+  "publicReview.rolesGuide.current.collector.description":
+    "Can mint with a valid signed permission, bid, withdraw, transfer, or burn when the current rules allow it.",
+  "publicReview.rolesGuide.current.signer.title": "Stream sale signer",
+  "publicReview.rolesGuide.current.signer.description":
+    "Approves one exact fixed-price mint or auction. Changing the signer also changes the signer version, so old permissions stop working.",
+  "publicReview.rolesGuide.current.admins.title": "Owner and administrators",
+  "publicReview.rolesGuide.current.admins.description":
+    "The owner assigns roles. A global administrator has broad power. A function administrator is limited to one named action on one contract.",
+  "publicReview.rolesGuide.current.pause.title":
+    "Pause and restart administrators",
+  "publicReview.rolesGuide.current.pause.description":
+    "A pause guardian can stop configured operations. A separate administrator decides when to restart them.",
+  "publicReview.rolesGuide.current.anyone.title": "Anyone",
+  "publicReview.rolesGuide.current.anyone.description":
+    "Anyone may settle an auction after it ends. The winner and payment amounts come from contract state.",
+  "publicReview.rolesGuide.inactive.heading":
+    "Built but not active in the current path",
+  "publicReview.rolesGuide.inactive.description":
+    "These roles exist in source or selected integrations. They are not all used by the current signed sale path.",
+  "publicReview.rolesGuide.inactive.mint.title": "Mint manager",
+  "publicReview.rolesGuide.inactive.mint.description":
+    "Can prepare and execute mints under policy and counter checks. The current signed Drop still uses the older mint route.",
+  "publicReview.rolesGuide.inactive.revenue.title":
+    "Revenue resolver and split contracts",
+  "publicReview.rolesGuide.inactive.revenue.description":
+    "Connect to selected parts of Stream. Current Drops and auctions still record sale credits inside their own contracts.",
+  "publicReview.rolesGuide.inactive.governance.title": "Governance roles",
+  "publicReview.rolesGuide.inactive.governance.description":
+    "Separate roles can propose, cancel, veto, and execute delayed changes. Final accounts, targets, delays, and value limits are not yet bound.",
+  "publicReview.rolesGuide.inactive.records.title": "Record writers",
+  "publicReview.rolesGuide.inactive.records.description":
+    "Code limits who may add each kind of record. Final record types, permission providers, and grants are not yet bound.",
+  "publicReview.rolesGuide.inactive.modules.title":
+    "Module and successor roles",
+  "publicReview.rolesGuide.inactive.modules.description":
+    "Code can record a replacement service while keeping the older service visible. The final module map and handover evidence are not yet proven.",
+  "publicReview.rolesGuide.inactive.randomness.title": "Randomness roles",
+  "publicReview.rolesGuide.inactive.randomness.description":
+    "Code separates randomness setup and delivery. Live provider setup, funding, monitoring, and recovery remain unproven.",
+  "publicReview.rolesGuide.future.heading": "Planned or still open",
+  "publicReview.rolesGuide.future.description":
+    "The ADRs describe these ideas. They must not be shown as protection that Stream already provides.",
+  "publicReview.rolesGuide.future.revenueAdapter.title":
+    "Revenue validation adapter",
+  "publicReview.rolesGuide.future.revenueAdapter.description":
+    "The design is accepted, but the complete adapter and matching resolver are not implemented in this snapshot.",
+  "publicReview.rolesGuide.future.artistAdapter.title":
+    "Artist identity adapter",
+  "publicReview.rolesGuide.future.artistAdapter.description":
+    "ADR 0022 proposes this check. It is not approved or implemented as a current protection.",
+  "publicReview.rolesGuide.future.artistRecovery.title":
+    "Artist authority and recovery",
+  "publicReview.rolesGuide.future.artistRecovery.description":
+    "Delegates, guardians, estates, sanctions, disputes, and recovery do not yet have a complete design.",
+  "publicReview.rolesGuide.future.finalityRecovery.title":
+    "Finality recovery record",
+  "publicReview.rolesGuide.future.finalityRecovery.description":
+    "ADR 0020 proposes a visible recovery record that keeps the original finality history. It is not implemented.",
+  "publicReview.rolesGuide.outside.heading":
+    "Responsibilities outside the contracts",
+  "publicReview.rolesGuide.outside.description":
+    "Smart contracts cannot control every part of Stream. These duties still depend on people and services.",
+  "publicReview.rolesGuide.outside.community.title":
+    "Community decisions and TDH",
+  "publicReview.rolesGuide.outside.community.description":
+    "Curation and TDH decisions happen outside the contracts before a sale permission is signed.",
+  "publicReview.rolesGuide.outside.signing.title": "Signer security",
+  "publicReview.rolesGuide.outside.signing.description":
+    "People and services must protect signing keys, rotate them after a problem, and watch for unexpected permissions.",
+  "publicReview.rolesGuide.outside.services.title":
+    "Storage and public services",
+  "publicReview.rolesGuide.outside.services.description":
+    "Artwork storage, websites, browsers, RPC nodes, and indexers must remain available and show the correct data.",
+  "publicReview.rolesGuide.outside.operations.title":
+    "Deployment and incident response",
+  "publicReview.rolesGuide.outside.operations.description":
+    "Safe owners, deployment teams, and monitors need clear owners, recovery steps, and public evidence.",
+  "publicReview.rolesGuide.outside.marketplaces.title": "Marketplaces",
+  "publicReview.rolesGuide.outside.marketplaces.description":
+    "Stream can report royalty information. A marketplace still decides whether to pay it.",
+  "publicReview.rolesGuide.risks.heading": "Main risks",
+  "publicReview.rolesGuide.risks.description":
+    "A clear role name does not guarantee that the power is safe. Review the full action and its limits.",
+  "publicReview.rolesGuide.risks.admin.title": "A role is too broad",
+  "publicReview.rolesGuide.risks.admin.description":
+    "A global or function role may reach more actions than people expect.",
+  "publicReview.rolesGuide.risks.signer.title": "A signer is stolen or wrong",
+  "publicReview.rolesGuide.risks.signer.description":
+    "A stolen key or bad signing service may approve a sale the community did not choose.",
+  "publicReview.rolesGuide.risks.pause.title": "A pause blocks safe exits",
+  "publicReview.rolesGuide.risks.pause.description":
+    "An emergency stop may also block refunds, withdrawals, settlement, or unfinished mints.",
+  "publicReview.rolesGuide.risks.artist.title":
+    "The artist approves less than the page suggests",
+  "publicReview.rolesGuide.risks.artist.description":
+    "The product may imply broad approval even when the signature covers only a small set of facts.",
+  "publicReview.rolesGuide.risks.status.title":
+    "Source code is mistaken for an active protection",
+  "publicReview.rolesGuide.risks.status.description":
+    "A control may exist in Solidity without being connected, configured, or ready for release.",
+  "publicReview.rolesGuide.risks.successor.title":
+    "Old and new services overlap",
+  "publicReview.rolesGuide.risks.successor.description":
+    "A replacement service may accept the same permission or responsibility as the service it replaces.",
+  "publicReview.rolesGuide.questions.heading": "Questions for reviewers",
+  "publicReview.rolesGuide.questions.description":
+    "These decisions should be clear before deployment.",
+  "publicReview.rolesGuide.questions.global":
+    "Which broad administrator roles should end before launch?",
+  "publicReview.rolesGuide.questions.delay":
+    "Which changes need a public waiting period?",
+  "publicReview.rolesGuide.questions.artist":
+    "Which actions need approval from both the artist and Stream operators?",
+  "publicReview.rolesGuide.questions.pause":
+    "Can collectors still settle, withdraw, or receive refunds during a pause?",
+  "publicReview.rolesGuide.questions.successor":
+    "What public evidence is required before a replacement service becomes active?",
+  "publicReview.rolesGuide.evidence.heading": "Check the detailed evidence",
+  "publicReview.rolesGuide.evidence.description":
+    "Use the development and governance pages to check the exact code status, open risks, and release evidence.",
+  "publicReview.rolesGuide.evidence.readPage": "Read {page}",
+} as const;
