+++ i18n/messages/public-review-community.ts
@@ -1,0 +1,224 @@
+export const PUBLIC_REVIEW_COMMUNITY_MESSAGES = {
+  "publicReview.community.editorial.heading":
+    "How to participate in the community review",
+  "publicReview.community.editorial.intro.purpose":
+    "Use this review to report an unclear promise, product problem, code bug, security risk, or missing proof before Stream's permanent rules are fixed.",
+  "publicReview.community.editorial.intro.noCode":
+    "You do not need Solidity knowledge. Explain what you expected, what you found, and why it matters.",
+  "publicReview.community.editorial.intro.source":
+    "This page covers review version `{reviewVersion}` and source commit {sourceLink}. Check both before commenting.",
+  "publicReview.community.editorial.people.heading": "Who should comment",
+  "publicReview.community.editorial.people.intro":
+    "Different experience reveals different problems:",
+  "publicReview.community.editorial.people.artists":
+    "Artists can review consent, creative control, preservation, recovery, and finality.",
+  "publicReview.community.editorial.people.collectors":
+    "Collectors can review ownership, artwork access, sale terms, and long-term promises.",
+  "publicReview.community.editorial.people.engineers":
+    "Contract, security, storage, browser, metadata, and infrastructure engineers can test technical and operational risks.",
+  "publicReview.community.editorial.people.product":
+    "Product, frontend, accessibility, and localization reviewers can find confusing or excluding workflows.",
+  "publicReview.community.editorial.people.community":
+    "Community members can ask whether Stream's powers and permanent choices match 6529's values.",
+  "publicReview.community.editorial.people.outro":
+    "Solidity knowledge is optional. An unclear promise or unfair workflow is useful feedback.",
+  "publicReview.community.editorial.scope.heading": "What to review",
+  "publicReview.community.editorial.scope.intro": "Useful feedback can cover:",
+  "publicReview.community.editorial.scope.artist":
+    "Artist identity, consent, collaborators, recovery, and estates.",
+  "publicReview.community.editorial.scope.sales":
+    "Collections, minting, sales, auctions, payments, splits, and royalties.",
+  "publicReview.community.editorial.scope.services":
+    "Randomness, metadata, storage, browsers, marketplaces, and indexers.",
+  "publicReview.community.editorial.scope.artwork":
+    "Artwork changes, preservation, collection freeze, and separate artwork finality.",
+  "publicReview.community.editorial.scope.governance":
+    "Roles, governance, emergency pauses, contract replacement, and powers that end permanently.",
+  "publicReview.community.editorial.scope.evidence":
+    "Threats, tests, launch setup, outside services, audit, and release evidence.",
+  "publicReview.community.editorial.scope.question":
+    "For every mechanism, ask what it protects, whether its power is limited, and whether a simpler design would work.",
+  "publicReview.community.editorial.safety.heading":
+    "Before reporting a security issue",
+  "publicReview.community.editorial.safety.policy":
+    "The current public-review policy allows possible Stream vulnerabilities to be reported in the public review Wave.",
+  "publicReview.community.editorial.safety.neverShare": "Never include:",
+  "publicReview.community.editorial.safety.credentials":
+    "Private keys, seed phrases, passwords, cookies, credentials, or API keys.",
+  "publicReview.community.editorial.safety.personal":
+    "Personal information that is not needed to explain the issue.",
+  "publicReview.community.editorial.safety.unrelated":
+    "Secrets or private data from another system.",
+  "publicReview.community.editorial.safety.attack":
+    "Instructions or data for attacking a live system outside this review.",
+  "publicReview.community.editorial.safety.otherProtocol":
+    "If the same technique affects another live protocol, keep the Stream report focused on Stream and coordinate the other disclosure separately.",
+  "publicReview.community.editorial.submit.heading": "How to submit feedback",
+  "publicReview.community.editorial.submit.intro":
+    "Start from the review page that best matches the issue:",
+  "publicReview.community.editorial.submit.page":
+    "Open the feedback panel on that page.",
+  "publicReview.community.editorial.submit.section":
+    "Choose the relevant section, or choose the whole page.",
+  "publicReview.community.editorial.submit.classify":
+    "Choose the closest feedback type and likely impact.",
+  "publicReview.community.editorial.submit.explain":
+    "Explain what happened and add evidence when you have it.",
+  "publicReview.community.editorial.submit.send":
+    "Submit the report. It becomes part of the public Stream review Wave.",
+  "publicReview.community.editorial.submit.crossModule":
+    "If the issue spans several contracts or pages, submit it from the main topic and link the other evidence.",
+  "publicReview.community.editorial.submit.short":
+    "Short reports are welcome. One precise question can reveal a missing rule.",
+  "publicReview.community.editorial.type.heading":
+    "Choose the closest feedback type",
+  "publicReview.community.editorial.type.intro":
+    "Choose the main reason for the report:",
+  "publicReview.community.editorial.type.question":
+    "Question — something is unclear or needs an official answer.",
+  "publicReview.community.editorial.type.documentation":
+    "Documentation — an explanation is wrong, incomplete, or hard to understand.",
+  "publicReview.community.editorial.type.artist":
+    "Artist workflow — the design does not fit a real artist's process.",
+  "publicReview.community.editorial.type.product":
+    "Product or UX — an action, result, or recovery path is confusing.",
+  "publicReview.community.editorial.type.protocol":
+    "Protocol design — a role, rule, lifecycle, or permanent choice should change.",
+  "publicReview.community.editorial.type.bug":
+    "Implementation bug — the code does not match the stated behavior.",
+  "publicReview.community.editorial.type.security":
+    "Possible security vulnerability — a path may allow theft, an unauthorized change, a permanent lock, manipulation, or takeover.",
+  "publicReview.community.editorial.type.testing":
+    "Testing or evidence gap — an important promise lacks a repeatable test or real-world proof.",
+  "publicReview.community.editorial.type.accessibility":
+    "Accessibility or localization — the review or product excludes a group of people.",
+  "publicReview.community.editorial.type.outro":
+    "Choose one main type. Use the report to explain where the issue crosses into other areas.",
+  "publicReview.community.editorial.impact.heading": "Choose the likely impact",
+  "publicReview.community.editorial.impact.intro":
+    "This is your first estimate, not a final decision:",
+  "publicReview.community.editorial.impact.critical":
+    "Critical — a credible risk of major asset loss, permanent unauthorized artwork change, or protocol-wide takeover.",
+  "publicReview.community.editorial.impact.high":
+    "High — a serious asset, authority, availability, or finality failure with a realistic path.",
+  "publicReview.community.editorial.impact.medium":
+    "Medium — a meaningful but narrower failure, or one with stronger conditions or a practical workaround.",
+  "publicReview.community.editorial.impact.low":
+    "Low — limited impact or an extra safety improvement.",
+  "publicReview.community.editorial.impact.informational":
+    "Informational — a clarification, convention, maintenance, or evidence improvement.",
+  "publicReview.community.editorial.impact.unknown":
+    "Pending assessment — use this when you are unsure about the impact.",
+  "publicReview.community.editorial.impact.outro":
+    "State what you know, what you assume, and what remains uncertain.",
+  "publicReview.community.editorial.report.heading": "Write a useful report",
+  "publicReview.community.editorial.report.intro":
+    "Answer as many of these questions as you can:",
+  "publicReview.community.editorial.report.expected": "What did you expect?",
+  "publicReview.community.editorial.report.observed":
+    "What did you find in the pinned code or review page?",
+  "publicReview.community.editorial.report.affected":
+    "Which person, collection, token, role, contract, field, or state is affected?",
+  "publicReview.community.editorial.report.preconditions":
+    "What must happen first, or what must an actor control?",
+  "publicReview.community.editorial.report.consequence":
+    "What could happen because of the issue?",
+  "publicReview.community.editorial.report.reproduction":
+    "Can you reproduce it with an example, call sequence, transaction, or test?",
+  "publicReview.community.editorial.report.fix":
+    "What change should be considered, and what remains uncertain?",
+  "publicReview.community.editorial.report.product":
+    "For a product or documentation issue, include the wording or screen state that caused the misunderstanding.",
+  "publicReview.community.editorial.report.artist":
+    "For an artist-workflow issue, describe the real studio, collaborator, signing, preservation, or estate process that is missing.",
+  "publicReview.community.editorial.technical.heading":
+    "Add technical evidence when needed",
+  "publicReview.community.editorial.technical.intro":
+    "This section is mainly for engineers and auditors. Use the generated Technical Reference when possible because it points to the pinned source.",
+  "publicReview.community.editorial.technical.reference":
+    "Link every important contract, function, or definition.",
+  "publicReview.community.editorial.technical.commit":
+    "If you use GitHub, link to the exact reviewed commit.",
+  "publicReview.community.editorial.technical.order":
+    "Describe the order of calls across contracts.",
+  "publicReview.community.editorial.technical.state":
+    "Explain which stored values change and which steps fail or remain.",
+  "publicReview.community.editorial.technical.test":
+    "Include the test, script, or settings needed to reproduce the issue.",
+  "publicReview.community.editorial.technical.paths":
+    "For a pause, replacement, successor, or finality bypass, include every other path that can create the same result.",
+  "publicReview.community.editorial.after.heading":
+    "What happens after submission",
+  "publicReview.community.editorial.after.new":
+    "Every new structured report appears in the review record with the state `NEW`.",
+  "publicReview.community.editorial.after.replies":
+    "Replies remain in the linked public Wave discussion.",
+  "publicReview.community.editorial.after.currentLimit":
+    "The current frontend does not yet record accepted, rejected, confirmed, or fixed states. Its only report state is `NEW`.",
+  "publicReview.community.editorial.after.future":
+    "A future response system should record who responded, the source change or decision, supporting tests, remaining risk, and the review version.",
+  "publicReview.community.editorial.record.heading":
+    "How the review record works",
+  "publicReview.community.editorial.record.intro":
+    "Each structured report stores four machine-readable fields:",
+  "publicReview.community.editorial.record.schema":
+    "`review_schema` — the format version used to read the report.",
+  "publicReview.community.editorial.record.type":
+    "`type` — the selected feedback type.",
+  "publicReview.community.editorial.record.severity":
+    "`severity` — the submitter's first impact estimate.",
+  "publicReview.community.editorial.record.context":
+    "`context` — the submission ID, review, version, page, optional section, and optional code reference.",
+  "publicReview.community.editorial.record.contextDetail":
+    "The code reference can pin a repository, commit, file, file hash, line range, contract, declaration, and selected-code hash.",
+  "publicReview.community.editorial.record.visible":
+    "The visible report uses normal human language. The stored fields support filtering and source checks.",
+  "publicReview.community.editorial.record.limit":
+    "The stored context does not include the frontend commit or the full wording shown on screen. By itself, it cannot prove the exact copy a person saw.",
+  "publicReview.community.editorial.record.filters":
+    "The current ledger can filter by feedback type, page, contract, impact, and the `NEW` state.",
+  "publicReview.community.editorial.versions.heading":
+    "When the source changes",
+  "publicReview.community.editorial.versions.new":
+    "A new contract candidate should receive a new immutable review bundle and version-specific routes.",
+  "publicReview.community.editorial.versions.old":
+    "Older versioned explanations, source links, and reports remain available.",
+  "publicReview.community.editorial.versions.manual":
+    "The current process compares review versions manually. It does not automatically decide whether an older report still applies.",
+  "publicReview.community.editorial.versions.carry":
+    "For each new candidate, record which earlier reports still apply, were fixed, or need another review.",
+  "publicReview.community.editorial.audit.heading":
+    "For auditors and review closeout",
+  "publicReview.community.editorial.audit.intro":
+    "The generated code inventory and public feedback record support an audit. They do not replace one.",
+  "publicReview.community.editorial.audit.currentHeading":
+    "What the current tools support",
+  "publicReview.community.editorial.audit.inventory":
+    "Reproduce the compiler-based code inventory.",
+  "publicReview.community.editorial.audit.reports":
+    "List reports that link to exact code.",
+  "publicReview.community.editorial.audit.filters":
+    "Filter reports by page, contract, type, impact, and `NEW` state.",
+  "publicReview.community.editorial.audit.export":
+    "Export the feedback records that have been loaded.",
+  "publicReview.community.editorial.audit.limit":
+    "Cross-version decisions and official response states are not automated. Auditors must compare them with published evidence.",
+  "publicReview.community.editorial.audit.closeoutHeading":
+    "What closeout should publish",
+  "publicReview.community.editorial.audit.candidate":
+    "The final candidate commit and Git tree.",
+  "publicReview.community.editorial.audit.versions":
+    "Every review version and available structural comparison.",
+  "publicReview.community.editorial.audit.risks":
+    "Unresolved questions and accepted risks.",
+  "publicReview.community.editorial.audit.fixes":
+    "Fix commits and repeatable test evidence.",
+  "publicReview.community.editorial.audit.auditReport":
+    "Audit reports and remediation status.",
+  "publicReview.community.editorial.audit.blockers":
+    "The final release-blocker checklist.",
+  "publicReview.community.editorial.audit.archive":
+    "A machine-readable archive of the review record.",
+  "publicReview.community.editorial.audit.outro":
+    "Independent experts must still audit the exact release candidate and launch configuration.",
+} as const;
