+++ i18n/messages/public-review.ts
@@ -1,3 +1,8 @@
+import { PUBLIC_REVIEW_ARTISTS_AND_ROLES_MESSAGES } from "@/i18n/messages/public-review-artists-and-roles";
+import { PUBLIC_REVIEW_ARTWORK_LIFECYCLE_MESSAGES } from "@/i18n/messages/public-review-artwork-lifecycle";
+import { PUBLIC_REVIEW_COMMUNITY_MESSAGES } from "@/i18n/messages/public-review-community";
+import { PUBLIC_REVIEW_DEVELOPMENT_MESSAGES } from "@/i18n/messages/public-review-development";
+
 export const PUBLIC_REVIEW_MESSAGES = {
   "navigation.nfts.theMemes": "The Memes",
   "navigation.nfts.gradient": "6529 Gradient",
@@ -164,1913 +169,10 @@ export const PUBLIC_REVIEW_MESSAGES = {
   "publicReview.overviewGuide.audiences.auditors.description":
     "Start with known limits, test evidence, and the risks that still block release.",
   "publicReview.overviewGuide.startPage": "Start with {page}",
-  "publicReview.forArtistsGuide.artwork.heading":
-    "Understand the plan before you approve it",
-  "publicReview.forArtistsGuide.artwork.description":
-    "Before launch, check the artwork, edition size, sale, payments, and who can still make changes. Stream records the exact version you approve.",
-  "publicReview.forArtistsGuide.artwork.identity.title":
-    "What artwork are you publishing?",
-  "publicReview.forArtistsGuide.artwork.identity.description":
-    "Choose the files, scripts, and other parts the artwork needs. Stream records the artist, the artwork, and the tokens created from it.",
-  "publicReview.forArtistsGuide.artwork.editions.title":
-    "Is it unique or an edition?",
-  "publicReview.forArtistsGuide.artwork.editions.description":
-    "Choose whether there is one token or an edition, how many tokens can exist, and how each token is created. If the work uses randomness, Stream saves the result for each token so it can be recreated later.",
-  "publicReview.forArtistsGuide.artwork.sales.title":
-    "How can collectors get it?",
-  "publicReview.forArtistsGuide.artwork.sales.description":
-    "The plan shows whether collectors mint at a fixed price or bid in an auction. It also shows the sale rules and who can take part.",
-  "publicReview.forArtistsGuide.artwork.payments.title":
-    "Where does the money go?",
-  "publicReview.forArtistsGuide.artwork.payments.description":
-    "Check the price, currency, and everyone who receives a share. Stream can state a royalty for later sales, but outside marketplaces may not pay it.",
-  "publicReview.forArtistsGuide.artwork.approval.title":
-    "What are you approving?",
-  "publicReview.forArtistsGuide.artwork.approval.description":
-    "Your wallet approves one exact version of the plan. If important details change, you need to review and approve the new version.",
-  "publicReview.forArtistsGuide.artwork.control.title":
-    "What can change or become permanent?",
-  "publicReview.forArtistsGuide.artwork.control.description":
-    "Before finality, some files, supply, minting, and sale settings can still change. Stream keeps a record of those changes. The finality process is meant to close the remaining ways the artwork itself can change.",
-  "publicReview.forArtistsGuide.journey.heading":
-    "How your artwork moves through Stream",
-  "publicReview.forArtistsGuide.journey.description":
-    "Your artwork moves through six stages. Each stage has different choices, checks, and approvals.",
-  "publicReview.forArtistsGuide.journey.listLabel":
-    "Artwork journey stages",
-  "publicReview.forArtistsGuide.journey.prepare.title": "Prepare",
-  "publicReview.forArtistsGuide.journey.prepare.description":
-    "Add the artwork files and anything they need to work. Choose the edition size, sale method, credits, storage, and payment shares.",
-  "publicReview.forArtistsGuide.journey.review.title": "Review",
-  "publicReview.forArtistsGuide.journey.review.description":
-    "Check the complete plan. Make sure the artwork, sale, payments, and who can still make changes are correct.",
-  "publicReview.forArtistsGuide.journey.approve.title": "Approve",
-  "publicReview.forArtistsGuide.journey.approve.description":
-    "Use your wallet to approve one exact version of the plan. If an important detail changes, review and approve the new version.",
-  "publicReview.forArtistsGuide.journey.select.title": "Community selection",
-  "publicReview.forArtistsGuide.journey.select.description":
-    "The community reviews the work outside the contracts. If it is selected, a separate Stream wallet signs the exact mint or auction terms.",
-  "publicReview.forArtistsGuide.journey.launch.title": "Launch",
-  "publicReview.forArtistsGuide.journey.launch.description":
-    "Stream checks the signed permission. Collectors can then mint or bid using only those sale terms.",
-  "publicReview.forArtistsGuide.journey.finalize.title": "Finalize",
-  "publicReview.forArtistsGuide.journey.finalize.description":
-    "Finality has a waiting period. During this time, the artist and reviewers check the files, supply, and every remaining way the artwork could change. The final step is meant to close those remaining change paths.",
-  "publicReview.forArtistsGuide.journey.important":
-    "Important: Ending minting, freezing settings, recording preservation evidence, and finalizing the artwork are separate steps. Completing one does not complete the others.",
-  "publicReview.forArtistsGuide.approval.heading": "Before you approve",
-  "publicReview.forArtistsGuide.approval.description":
-    "Your wallet approval covers specific details, not every part of Stream. Check what the signature covers and review the wider artwork plan before signing.",
-  "publicReview.forArtistsGuide.approval.listLabel": "Approval checklist",
-  "publicReview.forArtistsGuide.approval.signing.title":
-    "What you are signing",
-  "publicReview.forArtistsGuide.approval.signing.description":
-    "Check the artwork, artist wallet, contract, network, and version.",
-  "publicReview.forArtistsGuide.approval.artworkSupply.title":
-    "Artwork and supply",
-  "publicReview.forArtistsGuide.approval.artworkSupply.description":
-    "Check every required file and the maximum number of tokens. Include every way more tokens can be created.",
-  "publicReview.forArtistsGuide.approval.salePayments.title":
-    "Sale and payments",
-  "publicReview.forArtistsGuide.approval.salePayments.description":
-    "Check the sale type, price, currency, refunds, payment recipients, and shares.",
-  "publicReview.forArtistsGuide.approval.powerFinality.title":
-    "Power and finality",
-  "publicReview.forArtistsGuide.approval.powerFinality.description":
-    "Check who can change, pause, or restart each part. Check what every permanent step locks.",
-  "publicReview.forArtistsGuide.approval.note":
-    "If a signed detail changes, the old approval stays with the old version. Review the new version before signing again.",
-  "publicReview.forArtistsGuide.changes.heading": "What can still change?",
-  "publicReview.forArtistsGuide.changes.description":
-    "Stream makes parts of your artwork permanent at different times. Ending minting, freezing settings, and artwork finality are separate steps.",
-  "publicReview.forArtistsGuide.changes.beforeLaunch.title":
-    "Before minting starts",
-  "publicReview.forArtistsGuide.changes.beforeLaunch.description":
-    "Review the plan. If something is wrong, do not approve it. A corrected version can be prepared for you.",
-  "publicReview.forArtistsGuide.changes.afterLaunch.title":
-    "After minting starts",
-  "publicReview.forArtistsGuide.changes.afterLaunch.description":
-    "Minted tokens, payments, and signed actions stay in the history. Some settings can still change. Important changes may need a new approval from you.",
-  "publicReview.forArtistsGuide.changes.afterFinality.title":
-    "After artwork finality",
-  "publicReview.forArtistsGuide.changes.afterFinality.description":
-    "Finality is designed to lock the artwork records it covers. Before this happens, a waiting period lets the guardian stop a suspicious finality action.",
-  "publicReview.forArtistsGuide.actors.heading":
-    "Who else can affect your artwork?",
-  "publicReview.forArtistsGuide.actors.description":
-    "Your approval matters, but other people and services also have a role. Before you sign, check what each one can do and where its power ends.",
-  "publicReview.forArtistsGuide.actors.community.title": "Community review",
-  "publicReview.forArtistsGuide.actors.community.description":
-    "The community reviews and selects the work outside the contracts. The contracts cannot tell whether this decision was fair or correct.",
-  "publicReview.forArtistsGuide.actors.signer.title":
-    "Stream signing wallet",
-  "publicReview.forArtistsGuide.actors.signer.description":
-    "If the work is selected, a separate Stream wallet signs the exact mint or auction terms. The contracts can check that signature.",
-  "publicReview.forArtistsGuide.actors.operators.title": "People who run Stream",
-  "publicReview.forArtistsGuide.actors.operators.description":
-    "The contracts give different roles different powers. Check who can change, pause, or restart each part of Stream.",
-  "publicReview.forArtistsGuide.actors.guardian.title": "Guardian",
-  "publicReview.forArtistsGuide.actors.guardian.description":
-    "The guardian is a safety brake. It can pause a specific part of Stream or stop a planned change during its waiting period. It cannot replace that change with different artwork or terms.",
-  "publicReview.forArtistsGuide.actors.services.title": "Outside services",
-  "publicReview.forArtistsGuide.actors.services.description":
-    "File storage, websites, services that provide random results, and marketplaces work outside the contracts. Stream can record links and evidence. It cannot keep a service online or force a marketplace to pay royalties.",
-  "publicReview.forArtistsGuide.sales.heading":
-    "How sales and payments work",
-  "publicReview.forArtistsGuide.sales.buy.title": "How collectors buy",
-  "publicReview.forArtistsGuide.sales.buy.description":
-    "Collectors pay the signed fixed price or bid in an auction. The current signed sale paths use ETH. Before launch, check the auction end time, cancellation rules, and refund rules.",
-  "publicReview.forArtistsGuide.sales.money.title":
-    "Where the money goes",
-  "publicReview.forArtistsGuide.sales.money.description":
-    "Check every person or wallet that receives money and their share. After the sale, Stream records how much each one can withdraw. Each person withdraws separately, so one failed payment does not block everyone else.",
-  "publicReview.forArtistsGuide.sales.royalties.title": "Later royalties",
-  "publicReview.forArtistsGuide.sales.royalties.description":
-    "Stream can show royalty information for later marketplace sales. The marketplace decides whether to pay it.",
-  "publicReview.forArtistsGuide.permanence.heading":
-    "Before your artwork becomes final",
-  "publicReview.forArtistsGuide.permanence.description":
-    "Finality is meant to lock the artwork records it covers. Ending minting, freezing settings, and adding preservation records are separate steps.",
-  "publicReview.forArtistsGuide.permanence.files.title":
-    "Check every required part",
-  "publicReview.forArtistsGuide.permanence.files.description":
-    "Make sure every file, script, font, and other required part can still be found and opened.",
-  "publicReview.forArtistsGuide.permanence.fingerprints.title":
-    "Check the recorded fingerprints",
-  "publicReview.forArtistsGuide.permanence.fingerprints.description":
-    "A file fingerprint is a code made from the file. A matching code shows that the file has not changed. It does not keep the file online.",
-  "publicReview.forArtistsGuide.permanence.record.title":
-    "Check the final record",
-  "publicReview.forArtistsGuide.permanence.record.description":
-    "Confirm the token supply, mint history, payments, and any saved randomness results.",
-  "publicReview.forArtistsGuide.permanence.delay.title":
-    "Use the waiting period",
-  "publicReview.forArtistsGuide.permanence.delay.description":
-    "Finality has a delay so you and other reviewers can find mistakes. A guardian can stop the action during this time, but cannot replace it with different artwork or terms.",
-  "publicReview.forArtistsGuide.permanence.reviewStatus.title":
-    "What is still under review",
-  "publicReview.forArtistsGuide.permanence.reviewStatus.description":
-    "Stream is meant to close every approved path that can change the artwork. The current review has not yet proved that every path is closed. Check the technical evidence before relying on finality.",
-  "publicReview.forArtistsGuide.nextStep.heading": "Your next step",
-  "publicReview.forArtistsGuide.nextStep.description":
-    "Review the exact approval package. If it matches the plan you accept, sign it. If anything is missing or unclear, ask for a corrected version.",
-  "publicReview.forArtistsGuide.readPage": "Read {page}",
-  "publicReview.forArtistsGuide.evidence.heading": "Read the contract details",
-  "publicReview.forArtistsGuide.evidence.description":
-    "The short sections below show what is in the reviewed code, what comes from accepted design decisions, and what is still proposed.",
-  "publicReview.forArtistsGuide.evidence.action":
-    "Continue to the contract details",
-  "publicReview.forArtistsDetails.heading": "How to read the details",
-  "publicReview.forArtistsDetails.description":
-    "The labels show what kind of evidence each section uses. A section can use more than one.",
-  "publicReview.forArtistsDetails.basis.code.label": "Reviewed code",
-  "publicReview.forArtistsDetails.basis.code.description":
-    "Includes behavior present in the pinned contract snapshot.",
-  "publicReview.forArtistsDetails.basis.accepted.label": "Accepted design",
-  "publicReview.forArtistsDetails.basis.accepted.description":
-    "Includes decisions approved in an ADR. They may not be complete in the pinned code.",
-  "publicReview.forArtistsDetails.basis.proposed.label": "Still proposed",
-  "publicReview.forArtistsDetails.basis.proposed.description":
-    "Includes ideas that are not approved or safe to rely on yet.",
-  "publicReview.forArtistsDetails.identity.title":
-    "Your collection has a durable identity",
-  "publicReview.forArtistsDetails.identity.intro":
-    "Stream is designed around one shared ERC-721 Core contract. It gives every token a global ID and a serial number inside its collection.",
-  "publicReview.forArtistsDetails.identity.point.artist":
-    "The Core records the collection and its artist wallet.",
-  "publicReview.forArtistsDetails.identity.point.burn":
-    "The collection link stays in the contract's history after a token is burned.",
-  "publicReview.forArtistsDetails.identity.point.modules":
-    "Sale, rendering, and randomness contracts can change without changing this Core identity.",
-  "publicReview.forArtistsDetails.identity.outro":
-    "Stream is not deployed yet. This describes the reviewed design, not a live guarantee.",
-  "publicReview.forArtistsDetails.approval.title":
-    "Approving a specific collection state",
-  "publicReview.forArtistsDetails.approval.intro":
-    "The current artist signature approves one exact collection state. It also binds the network and Core contract.",
-  "publicReview.forArtistsDetails.approval.point.artist": "Artist wallet",
-  "publicReview.forArtistsDetails.approval.point.freeze":
-    "Collection freeze fingerprint",
-  "publicReview.forArtistsDetails.approval.point.purchases":
-    "Maximum collection purchases",
-  "publicReview.forArtistsDetails.approval.point.supply":
-    "Total collection supply",
-  "publicReview.forArtistsDetails.approval.point.delay":
-    "Delay before final supply",
-  "publicReview.forArtistsDetails.approval.outro":
-    "If one value changes, the old signature stays with the old state. This signature does not directly approve every file, sale rule, payment share, or finality action. Check those separately.",
-  "publicReview.forArtistsDetails.scope.title":
-    "The scope of artist approval",
-  "publicReview.forArtistsDetails.scope.intro":
-    "Today's signature is narrow. The wider design is still deciding which other actions need a fresh artist signature.",
-  "publicReview.forArtistsDetails.scope.point.files":
-    "Artwork files and manifests",
-  "publicReview.forArtistsDetails.scope.point.supply": "Final supply",
-  "publicReview.forArtistsDetails.scope.point.sale":
-    "Sale and payment rules",
-  "publicReview.forArtistsDetails.scope.point.randomness":
-    "Randomness provider",
-  "publicReview.forArtistsDetails.scope.point.freeze": "Core freeze",
-  "publicReview.forArtistsDetails.scope.point.finality": "Artwork finality",
-  "publicReview.forArtistsDetails.scope.point.recovery":
-    "Recovery that changes what viewers receive",
-  "publicReview.forArtistsDetails.scope.outro":
-    "These are open design questions, not current guarantees. A guardian may stop a planned action, but should not be able to replace it with different artwork.",
-  "publicReview.forArtistsDetails.statements.title":
-    "Statements made in the artist's name",
-  "publicReview.forArtistsDetails.statements.intro":
-    "Stream separates records by who is allowed to speak: artist, owner, institution, independent reviewer, and other roles.",
-  "publicReview.forArtistsDetails.statements.rules":
-    "The reviewed code limits who can submit protected record types. An ordinary admin cannot automatically submit an artist statement.",
-  "publicReview.forArtistsDetails.statements.point.author":
-    "Who submitted the statement",
-  "publicReview.forArtistsDetails.statements.point.authority":
-    "Which wallet or authority allowed it",
-  "publicReview.forArtistsDetails.statements.point.subject":
-    "Which collection or token it concerns",
-  "publicReview.forArtistsDetails.statements.point.change":
-    "Whether it can be revised, locked, or replaced by a newer record",
-  "publicReview.forArtistsDetails.statements.outro":
-    "A record can prove who submitted a claim. It cannot prove that the claim is true. Rights, archive, and provenance claims still need outside evidence.",
-  "publicReview.forArtistsDetails.files.title":
-    "Artwork files, scripts, and token data",
-  "publicReview.forArtistsDetails.files.intro":
-    "The artwork may depend on more than one image. It can include scripts, fonts, libraries, token data, and outside files.",
-  "publicReview.forArtistsDetails.files.check":
-    "Before approval or finality, ask for one clear file list. It should show:",
-  "publicReview.forArtistsDetails.files.point.parts":
-    "Every required file and script",
-  "publicReview.forArtistsDetails.files.point.order":
-    "The order in which scripts run",
-  "publicReview.forArtistsDetails.files.point.versions":
-    "The version of each dependency",
-  "publicReview.forArtistsDetails.files.point.locations":
-    "Where each file can be found",
-  "publicReview.forArtistsDetails.files.point.hashes":
-    "A fingerprint for the exact bytes",
-  "publicReview.forArtistsDetails.files.point.rebuild":
-    "Instructions for recreating the work",
-  "publicReview.forArtistsDetails.files.outro":
-    "A matching fingerprint shows that a file has not changed. It does not keep the file available.",
-  "publicReview.forArtistsDetails.editions.title":
-    "One-of-ones and editions",
-  "publicReview.forArtistsDetails.editions.intro":
-    "The Core can hold a one-of-one or an edition. Check:",
-  "publicReview.forArtistsDetails.editions.point.maximum":
-    "The maximum number of tokens",
-  "publicReview.forArtistsDetails.editions.point.minted":
-    "How many tokens have ever been minted",
-  "publicReview.forArtistsDetails.editions.point.live":
-    "How many live tokens remain after burns",
-  "publicReview.forArtistsDetails.editions.point.paths":
-    "Every remaining way to mint",
-  "publicReview.forArtistsDetails.editions.point.control":
-    "Who can change supply or mint rules",
-  "publicReview.forArtistsDetails.editions.point.close":
-    "What closes minting for good",
-  "publicReview.forArtistsDetails.editions.outro":
-    "Burning a token removes its ownership record. It does not erase mint history or restore the used mint allowance.",
-  "publicReview.forArtistsDetails.mint.title": "Choosing who can mint",
-  "publicReview.forArtistsDetails.mint.intro":
-    "Artwork identity and mint rules are separate. Mint rules can include times, supply limits, wallet limits, eligibility checks, and an approved signer.",
-  "publicReview.forArtistsDetails.mint.authorization":
-    "In the current signed Drop path, one authorization can mint only one token. An edition needs several authorizations or another approved mint path.",
-  "publicReview.forArtistsDetails.mint.outro":
-    "Before launch, check who can authorize a mint and how used, expired, or cancelled permissions are blocked.",
-  "publicReview.forArtistsDetails.curation.title": "Curation and TDH",
-  "publicReview.forArtistsDetails.curation.intro":
-    "Community curation and TDH happen outside the contracts.",
-  "publicReview.forArtistsDetails.curation.signature":
-    "If a work is selected, a Stream signing wallet signs the exact mint or auction terms. The contract checks that signature.",
-  "publicReview.forArtistsDetails.curation.limit":
-    "The contract can check that the signed terms were not changed. It cannot check whether the community process was fair or the selection was correct.",
-  "publicReview.forArtistsDetails.sales.title":
-    "Fixed-price sales and auctions",
-  "publicReview.forArtistsDetails.sales.intro":
-    "The reviewed sale paths use ETH. Before launch, check:",
-  "publicReview.forArtistsDetails.sales.point.fixed":
-    "Fixed price: token, price, payer, recipient, deadline, payment shares, and cancellation state",
-  "publicReview.forArtistsDetails.sales.point.auction":
-    "Auction: reserve, end time, extension rule, token custody, refunds, cancellation boundary, and no-bid result",
-  "publicReview.forArtistsDetails.sales.point.cancel":
-    "An auction can be cancelled only before its first valid bid and before it ends",
-  "publicReview.forArtistsDetails.sales.outro":
-    "The contracts record proceeds and refunds as credits. Each person withdraws their own credit later. One failed withdrawal does not block the sale.",
-  "publicReview.forArtistsDetails.revenue.title":
-    "Revenue, collaborators, and royalties",
-  "publicReview.forArtistsDetails.revenue.intro":
-    "Primary-sale money can be shared among the artist, collaborators, curator, protocol, and other recipients.",
-  "publicReview.forArtistsDetails.revenue.paths":
-    "The reviewed code has native ETH credit paths for Drops and Auctions. It also has a separate revenue resolver and split-wallet foundation. These are different payment paths.",
-  "publicReview.forArtistsDetails.revenue.point.people":
-    "Every recipient and share",
-  "publicReview.forArtistsDetails.revenue.point.profile":
-    "Which payment profile applies",
-  "publicReview.forArtistsDetails.revenue.point.change":
-    "Who can change the profile and when",
-  "publicReview.forArtistsDetails.revenue.point.rounding":
-    "How rounding, remaining funds, and withdrawals work",
-  "publicReview.forArtistsDetails.revenue.point.royalty":
-    "The royalty receiver and rate",
-  "publicReview.forArtistsDetails.revenue.outro":
-    "Stream can publish ERC-2981 royalty information. Marketplaces choose whether to pay it.",
-  "publicReview.forArtistsDetails.randomness.title": "Randomness",
-  "publicReview.forArtistsDetails.randomness.intro":
-    "If the artwork uses randomness, check the provider and failure plan before minting.",
-  "publicReview.forArtistsDetails.randomness.point.state":
-    "The reviewed code records whether a request is waiting, complete, stale, or received but not fully processed",
-  "publicReview.forArtistsDetails.randomness.point.result":
-    "It stores a fingerprint of the provider result and the seed made from it",
-  "publicReview.forArtistsDetails.randomness.point.retry":
-    "If local processing fails, a retry uses the same saved seed. It does not ask for a new random result",
-  "publicReview.forArtistsDetails.randomness.point.control":
-    "Check who can change the provider, what happens to unfinished requests, and who pays provider fees",
-  "publicReview.forArtistsDetails.freeze.title": "Freezing the work",
-  "publicReview.forArtistsDetails.freeze.intro":
-    "Four steps are easy to confuse:",
-  "publicReview.forArtistsDetails.freeze.point.supply":
-    "Final supply: no more tokens can be minted",
-  "publicReview.forArtistsDetails.freeze.point.core":
-    "Core freeze: locks the collection fields covered by the Core contract",
-  "publicReview.forArtistsDetails.freeze.point.preservation":
-    "Preservation records: save evidence about the files needed to understand or recreate the work",
-  "publicReview.forArtistsDetails.freeze.point.finality":
-    "Artwork finality: a delayed step intended to close the remaining ways the artwork can change",
-  "publicReview.forArtistsDetails.freeze.separate":
-    "These are separate. Completing one does not complete the others.",
-  "publicReview.forArtistsDetails.freeze.check":
-    "Before freeze or finality, check the files, fingerprints, supply, randomness, payments, open mint and sale paths, and who can still make changes.",
-  "publicReview.forArtistsDetails.freeze.limit":
-    "Accepted ADRs describe stronger freeze and finality goals. The current public review has not yet proved that every artwork-changing path is closed. Do not rely on finality until the technical review is complete.",
-  "publicReview.forArtistsDetails.lifetime.title":
-    "Collaborators, delegation, recovery, and estates",
-  "publicReview.forArtistsDetails.lifetime.intro":
-    "The current artist approval is not a complete lifetime authority system. A future system may need:",
-  "publicReview.forArtistsDetails.lifetime.point.collaborators":
-    "Collaborators with narrow roles",
-  "publicReview.forArtistsDetails.lifetime.point.delegation":
-    "Delegated signing powers",
-  "publicReview.forArtistsDetails.lifetime.point.recovery":
-    "Wallet recovery and key changes",
-  "publicReview.forArtistsDetails.lifetime.point.estates":
-    "Estate instructions and successor roles",
-  "publicReview.forArtistsDetails.lifetime.point.disputes":
-    "Rules for disputes and sanctions",
-  "publicReview.forArtistsDetails.lifetime.status":
-    "ADR 0022's registry and helper design is still proposed. It is not approved or implemented.",
-  "publicReview.forArtistsDetails.lifetime.question":
-    "The key question: can someone recover access without gaining power to replace the artwork or rewrite its history?",
-  "publicReview.forArtistsDetails.design.title": "Design position",
-  "publicReview.forArtistsDetails.design.intro":
-    "Artists should never have to sign a hash they cannot understand. Before a wallet asks for a signature, the screen should show:",
-  "publicReview.forArtistsDetails.design.point.work":
-    "The exact artwork and version",
-  "publicReview.forArtistsDetails.design.point.files":
-    "The files and token supply",
-  "publicReview.forArtistsDetails.design.point.sale":
-    "The sale and payment rules",
-  "publicReview.forArtistsDetails.design.point.randomness":
-    "The randomness provider",
-  "publicReview.forArtistsDetails.design.point.power":
-    "Who can change what",
-  "publicReview.forArtistsDetails.design.point.permanent":
-    "What becomes permanent",
-  "publicReview.forArtistsDetails.design.outro":
-    "The goal is simple: every important choice should be visible before it becomes final.",
-  "publicReview.forArtistsDetails.questions.title": "Questions for artists",
-  "publicReview.forArtistsDetails.questions.intro":
-    "These decisions are still worth discussing:",
-  "publicReview.forArtistsDetails.questions.point.signature":
-    "Which actions must need your signature?",
-  "publicReview.forArtistsDetails.questions.point.shared":
-    "Which actions should also need approval from the people who run Stream?",
-  "publicReview.forArtistsDetails.questions.point.collaborators":
-    "What should a collaborator be allowed to do?",
-  "publicReview.forArtistsDetails.questions.point.recovery":
-    "What recovery and estate process would you trust?",
-  "publicReview.forArtistsDetails.questions.point.preservation":
-    "What evidence would you need before freezing the work?",
-  "publicReview.forArtistsDetails.questions.point.payments":
-    "Which sale and payment settings should become permanent?",
-  "publicReview.forArtistsDetails.questions.point.finality":
-    "Would the finality process give you confidence, or add too much work?",
-  "publicReview.rolesGuide.status.heading": "Start with status",
-  "publicReview.rolesGuide.status.description":
-    "Stream is still being prepared. A role may exist in the code without being active in the current rehearsal. Check the label before treating a power as available.",
-  "publicReview.rolesGuide.status.working.title": "Working in rehearsal",
-  "publicReview.rolesGuide.status.working.description":
-    "Connected to the current rehearsal path.",
-  "publicReview.rolesGuide.status.connected.title":
-    "Connected for integration",
-  "publicReview.rolesGuide.status.connected.description":
-    "Connected to selected contracts, but not used by every current path.",
-  "publicReview.rolesGuide.status.source.title": "Built in source",
-  "publicReview.rolesGuide.status.source.description":
-    "The code exists. Final accounts, settings, and release wiring are not proven.",
-  "publicReview.rolesGuide.status.planned.title": "Planned",
-  "publicReview.rolesGuide.status.planned.description":
-    "The design is accepted. Complete code or wiring is still missing.",
-  "publicReview.rolesGuide.status.open.title": "Still open",
-  "publicReview.rolesGuide.status.open.description":
-    "No final design decision has been made.",
-  "publicReview.rolesGuide.current.heading":
-    "Working in the current rehearsal",
-  "publicReview.rolesGuide.current.description":
-    "These are the main powers used by the current signed Drop and auction path.",
-  "publicReview.rolesGuide.current.artist.title": "Artist",
-  "publicReview.rolesGuide.current.artist.description":
-    "Signs one exact collection state. The signature covers specific collection facts, not every later action.",
-  "publicReview.rolesGuide.current.collector.title": "Collector",
-  "publicReview.rolesGuide.current.collector.description":
-    "Can mint with a valid signed permission, bid, withdraw, transfer, or burn when the current rules allow it.",
-  "publicReview.rolesGuide.current.signer.title": "Stream sale signer",
-  "publicReview.rolesGuide.current.signer.description":
-    "Approves one exact fixed-price mint or auction. Changing the signer also changes the signer version, so old permissions stop working.",
-  "publicReview.rolesGuide.current.admins.title": "Owner and administrators",
-  "publicReview.rolesGuide.current.admins.description":
-    "The owner assigns roles. A global administrator has broad power. A function administrator is limited to one named action on one contract.",
-  "publicReview.rolesGuide.current.pause.title":
-    "Pause and restart administrators",
-  "publicReview.rolesGuide.current.pause.description":
-    "A pause guardian can stop configured operations. A separate administrator decides when to restart them.",
-  "publicReview.rolesGuide.current.anyone.title": "Anyone",
-  "publicReview.rolesGuide.current.anyone.description":
-    "Anyone may settle an auction after it ends. The winner and payment amounts come from contract state.",
-  "publicReview.rolesGuide.inactive.heading":
-    "Built but not active in the current path",
-  "publicReview.rolesGuide.inactive.description":
-    "These roles exist in source or selected integrations. They are not all used by the current signed sale path.",
-  "publicReview.rolesGuide.inactive.mint.title": "Mint manager",
-  "publicReview.rolesGuide.inactive.mint.description":
-    "Can prepare and execute mints under policy and counter checks. The current signed Drop still uses the older mint route.",
-  "publicReview.rolesGuide.inactive.revenue.title":
-    "Revenue resolver and split contracts",
-  "publicReview.rolesGuide.inactive.revenue.description":
-    "Connect to selected parts of Stream. Current Drops and auctions still record sale credits inside their own contracts.",
-  "publicReview.rolesGuide.inactive.governance.title": "Governance roles",
-  "publicReview.rolesGuide.inactive.governance.description":
-    "Separate roles can propose, cancel, veto, and execute delayed changes. Final accounts, targets, delays, and value limits are not yet bound.",
-  "publicReview.rolesGuide.inactive.records.title": "Record writers",
-  "publicReview.rolesGuide.inactive.records.description":
-    "Code limits who may add each kind of record. Final record types, permission providers, and grants are not yet bound.",
-  "publicReview.rolesGuide.inactive.modules.title":
-    "Module and successor roles",
-  "publicReview.rolesGuide.inactive.modules.description":
-    "Code can record a replacement service while keeping the older service visible. The final module map and handover evidence are not yet proven.",
-  "publicReview.rolesGuide.inactive.randomness.title": "Randomness roles",
-  "publicReview.rolesGuide.inactive.randomness.description":
-    "Code separates randomness setup and delivery. Live provider setup, funding, monitoring, and recovery remain unproven.",
-  "publicReview.rolesGuide.future.heading": "Planned or still open",
-  "publicReview.rolesGuide.future.description":
-    "The ADRs describe these ideas. They must not be shown as protection that Stream already provides.",
-  "publicReview.rolesGuide.future.revenueAdapter.title":
-    "Revenue validation adapter",
-  "publicReview.rolesGuide.future.revenueAdapter.description":
-    "The design is accepted, but the complete adapter and matching resolver are not implemented in this snapshot.",
-  "publicReview.rolesGuide.future.artistAdapter.title":
-    "Artist identity adapter",
-  "publicReview.rolesGuide.future.artistAdapter.description":
-    "ADR 0022 proposes this check. It is not approved or implemented as a current protection.",
-  "publicReview.rolesGuide.future.artistRecovery.title":
-    "Artist authority and recovery",
-  "publicReview.rolesGuide.future.artistRecovery.description":
-    "Delegates, guardians, estates, sanctions, disputes, and recovery do not yet have a complete design.",
-  "publicReview.rolesGuide.future.finalityRecovery.title":
-    "Finality recovery record",
-  "publicReview.rolesGuide.future.finalityRecovery.description":
-    "ADR 0020 proposes a visible recovery record that keeps the original finality history. It is not implemented.",
-  "publicReview.rolesGuide.outside.heading":
-    "Responsibilities outside the contracts",
-  "publicReview.rolesGuide.outside.description":
-    "Smart contracts cannot control every part of Stream. These duties still depend on people and services.",
-  "publicReview.rolesGuide.outside.community.title":
-    "Community decisions and TDH",
-  "publicReview.rolesGuide.outside.community.description":
-    "Curation and TDH decisions happen outside the contracts before a sale permission is signed.",
-  "publicReview.rolesGuide.outside.signing.title": "Signer security",
-  "publicReview.rolesGuide.outside.signing.description":
-    "People and services must protect signing keys, rotate them after a problem, and watch for unexpected permissions.",
-  "publicReview.rolesGuide.outside.services.title":
-    "Storage and public services",
-  "publicReview.rolesGuide.outside.services.description":
-    "Artwork storage, websites, browsers, RPC nodes, and indexers must remain available and show the correct data.",
-  "publicReview.rolesGuide.outside.operations.title":
-    "Deployment and incident response",
-  "publicReview.rolesGuide.outside.operations.description":
-    "Safe owners, deployment teams, and monitors need clear owners, recovery steps, and public evidence.",
-  "publicReview.rolesGuide.outside.marketplaces.title": "Marketplaces",
-  "publicReview.rolesGuide.outside.marketplaces.description":
-    "Stream can report royalty information. A marketplace still decides whether to pay it.",
-  "publicReview.rolesGuide.risks.heading": "Main risks",
-  "publicReview.rolesGuide.risks.description":
-    "A clear role name does not guarantee that the power is safe. Review the full action and its limits.",
-  "publicReview.rolesGuide.risks.admin.title": "A role is too broad",
-  "publicReview.rolesGuide.risks.admin.description":
-    "A global or function role may reach more actions than people expect.",
-  "publicReview.rolesGuide.risks.signer.title": "A signer is stolen or wrong",
-  "publicReview.rolesGuide.risks.signer.description":
-    "A stolen key or bad signing service may approve a sale the community did not choose.",
-  "publicReview.rolesGuide.risks.pause.title": "A pause blocks safe exits",
-  "publicReview.rolesGuide.risks.pause.description":
-    "An emergency stop may also block refunds, withdrawals, settlement, or unfinished mints.",
-  "publicReview.rolesGuide.risks.artist.title":
-    "The artist approves less than the page suggests",
-  "publicReview.rolesGuide.risks.artist.description":
-    "The product may imply broad approval even when the signature covers only a small set of facts.",
-  "publicReview.rolesGuide.risks.status.title":
-    "Source code is mistaken for an active protection",
-  "publicReview.rolesGuide.risks.status.description":
-    "A control may exist in Solidity without being connected, configured, or ready for release.",
-  "publicReview.rolesGuide.risks.successor.title":
-    "Old and new services overlap",
-  "publicReview.rolesGuide.risks.successor.description":
-    "A replacement service may accept the same permission or responsibility as the service it replaces.",
-  "publicReview.rolesGuide.questions.heading": "Questions for reviewers",
-  "publicReview.rolesGuide.questions.description":
-    "These decisions should be clear before deployment.",
-  "publicReview.rolesGuide.questions.global":
-    "Which broad administrator roles should end before launch?",
-  "publicReview.rolesGuide.questions.delay":
-    "Which changes need a public waiting period?",
-  "publicReview.rolesGuide.questions.artist":
-    "Which actions need approval from both the artist and Stream operators?",
-  "publicReview.rolesGuide.questions.pause":
-    "Can collectors still settle, withdraw, or receive refunds during a pause?",
-  "publicReview.rolesGuide.questions.successor":
-    "What public evidence is required before a replacement service becomes active?",
-  "publicReview.rolesGuide.evidence.heading": "Check the detailed evidence",
-  "publicReview.rolesGuide.evidence.description":
-    "Use the development and governance pages to check the exact code status, open risks, and release evidence.",
-  "publicReview.rolesGuide.evidence.readPage": "Read {page}",
-  "publicReview.pages.overviewNarrative.title": "Overview",
-  "publicReview.pages.overviewNarrative.summary":
-    "An introduction to Stream and its public review for artists, collectors, and the wider community.",
-  "publicReview.pages.artworkLifecycle.title": "Artwork Lifecycle",
-  "publicReview.pages.artworkLifecycle.summary":
-    "How an artwork moves from preparation through minting, preservation, and finality.",
-  "publicReview.pages.artworkLifecycle.currentSummary":
-    "This page is for artists, collectors, and people reviewing Stream. It explains how a Stream artwork moves from setup to its final state. No contract knowledge is needed.",
-  "publicReview.pages.artworkLifecycle.currentIntro":
-    "## The lifecycle in one minute\n\nA Stream artwork is built step by step.\n\nFirst, the collection gets a permanent identity in the Core. The Core is the shared home for all Stream tokens. Next, the artwork files, minting rules, payments, and other details are added. The artwork can then be minted or sold.\n\nLater, supply can close, key details can be locked, and preservation records can be added. Only then can the artwork reach its final state.\n\nMinted, sold, frozen, preserved, and final are different stages. This page explains what each stage means and why it matters.",
-  "publicReview.pages.artworkLifecycle.currentIdentitySection": `## 1. The collection gets a permanent identity
-
-### What happens
-
-Before anything is minted or sold, Stream gives the collection a permanent ID in the Core.
-
-The ID stays the same even if the collection later uses a different minting or sale tool.
-
-### Who creates it
-
-An account with permission to create collections starts it. Stream then records the artist, supply limit, purchase limit, and the wait before supply can become final.
-
-### What the current contract records
-
-The Core stores the collection ID and its basic information. Each token later receives:
-
-- one ID across all Stream tokens; and
-- one serial number inside its collection.
-
-**Why this matters:** The artwork keeps one clear identity even when the tools around it change.
-
-### Technical details
-
-The Core is the shared ERC-721 NFT contract. [\`StreamCore.createCollection\`](https://github.com/6529-Collections/6529Stream/blob/513bd7e079eafe109df6ae1ae21bfbca6fec6786/smart-contracts/StreamCore.sol#L336) creates the collection. [\`setCollectionData\`](https://github.com/6529-Collections/6529Stream/blob/513bd7e079eafe109df6ae1ae21bfbca6fec6786/smart-contracts/StreamCore.sol#L379) stores its artist and supply settings.`,
-  "publicReview.pages.artworkLifecycle.currentArtworkPackageSection": `## 2. The artwork package is prepared
-
-### What happens
-
-A Stream artwork is more than an image. Its package can include:
-
-- artwork files, such as images, animation, and code;
-- details, such as its name, traits, and collection;
-- supply, minting, and sale rules;
-- payment details;
-- randomness rules, if the artwork uses them; and
-- preservation and final-state records.
-
-### Who controls each part
-
-Different records can have different approved writers. For example, an artist controls an artist statement. An independent institution can add preservation evidence.
-
-### What stays permanent
-
-The token’s identity stays in the Core. Other tools, such as a display module or randomness service, may need a replacement years later.
-
-The current contracts keep these parts separate and restrict who may write each kind of record.
-
-**Why this matters:** A tool can be replaced without giving the artwork a new identity.
-
-### Technical details
-
-Display, sale, randomness, and preservation features can live in separate modules. Record-family checks keep artist, owner, institution, observer, rights, and archive records with their approved writers.`,
-  "publicReview.pages.artworkLifecycle.currentArtistApprovalSection": `## 3. The artist can sign the current setup
-
-### What happens
-
-The artist can sign one exact snapshot of the collection.
-
-That snapshot includes:
-
-- the artist’s wallet;
-- collection information, scripts, and dependencies;
-- current live supply and token metadata;
-- randomness settings; and
-- purchase, supply, and final-supply settings.
-
-A successful mint changes the live supply and token metadata. The old signature then stops matching the current snapshot.
-
-### Who decides
-
-Only the recorded artist can approve the snapshot. The artist can sign directly or provide a signed message that another account submits.
-
-### What the current contract does
-
-The contract stores the approved snapshot and can report whether it still matches the current state.
-
-This signature is evidence only. The minting paths do not check it. A missing or outdated signature does not pause or stop minting.
-
-### How this differs from the ADR design
-
-This snapshot signature is not the mint-policy approval required by the accepted design. That separate approval is explained in the next section.
-
-**Why this matters:** People can see what the artist signed without mistaking the signature for permission to mint.
-
-### Technical details
-
-The current snapshot signature uses EIP-712. It is tied to the chain and Core contract. It supports ordinary wallets and ERC-1271 smart-contract wallets. The snapshot includes the collection-freeze manifest hash, which changes when live supply or live token metadata changes.`,
-  "publicReview.pages.artworkLifecycle.currentDistributionSection": `## 4. The minting rules are chosen
-
-### What happens
-
-Before collectors can mint, rules are set for:
-
-- the type of release;
-- its opening and closing time;
-- who may mint;
-- the price; and
-- supply and wallet limits.
-
-### Who decides
-
-Accounts with the required Stream roles configure the current contracts.
-
-The accepted design adds an artist check. For an artist-bound collection, the artist must approve the mint policy or give someone limited signed permission to act for them.
-
-An ADR, or Architecture Decision Record, is an accepted design decision. It describes the target design, which may be ahead of the reviewed code.
-
-The artist permission must be checked before every mint. The same permission can cover later mints while the policy stays the same. A new permission is needed when the policy changes.
-
-The reviewed contracts do not yet enforce this artist-permission check.
-
-### What the current contracts check
-
-The reviewed code has two separate minting paths:
-
-- The signed-drop and auction path checks its own sale time and supply rules.
-- The manager path checks phases, approved executors, optional access gates, policy details, and usage limits.
-
-The paths do not share every check or counter. Each path must be reviewed on its own.
-
-**Why this matters:** Minting tools can change without changing the token’s permanent identity, but every path must enforce the right limits.
-
-### Technical details
-
-The signed-drop and auction path uses [\`StreamMinter\`](https://github.com/6529-Collections/6529Stream/blob/513bd7e079eafe109df6ae1ae21bfbca6fec6786/smart-contracts/StreamMinter.sol). The manager path uses [\`StreamMintManager\`](https://github.com/6529-Collections/6529Stream/blob/513bd7e079eafe109df6ae1ae21bfbca6fec6786/smart-contracts/StreamMintManager.sol) and [\`StreamMintLedger\`](https://github.com/6529-Collections/6529Stream/blob/513bd7e079eafe109df6ae1ae21bfbca6fec6786/smart-contracts/StreamMintLedger.sol). The Core keeps the token identity while these outside modules apply minting policy.`,
-  "publicReview.pages.artworkLifecycle.currentCurationSection": `## 5. The selected drop receives signed approval
-
-### What happens
-
-The community process applies its curation rules outside the blockchain. These rules may use TDH, which means Total Days Held. TDH measures how long eligible assets have been held.
-
-A service then prepares the exact sale details. An approved signer signs them. The signer is a wallet trusted to approve the result.
-
-### Who decides
-
-The community process chooses the artist, calculates TDH, and applies its fairness rules. The signer approves the exact action created from that result.
-
-The contract does not choose the artist, calculate TDH, or decide whether the result is fair.
-
-### What the current contract checks
-
-The signed approval fixes:
-
-- the blockchain and Stream contract;
-- the collection and artwork;
-- who pays and who receives the token;
-- the fixed price or starting auction terms;
-- when the approval expires; and
-- a unique ID that prevents reuse.
-
-Before the sale starts, the contract checks the signer and every signed detail. Changing the artwork, buyer, price, sale type, or deadline makes the approval invalid.
-
-In this path, one approval covers one token. After a successful use, it cannot be used again.
-
-**Why this matters:** The contract can only start the sale with the terms that were approved.`,
-  "publicReview.pages.artworkLifecycle.currentMintExecutionSection": `## 6. The mint completes fully or not at all
-
-### What happens
-
-The chosen minting path checks its rules and asks the Core to create the token.
-
-### What each path checks
-
-**Signed-drop and auction path**
-
-- The drop contract checks the signed approval, sale details, and any required payment.
-- The minter checks that minting is not paused, the minting window is open, and enough supply remains.
-- The Core applies its own collection and token checks.
-
-**Manager path**
-
-- The manager checks the active phase, approved executor, optional access rules, expected policy, and quantity.
-- The ledger records the limits used by that path.
-- The manager and Core prepare and complete the token in the same transaction.
-
-These paths do not use the same approval or counters.
-
-### What happens if a check fails
-
-All checks and changes happen in one blockchain transaction. If a check fails, the whole transaction is cancelled:
-
-- no token is created;
-- supply does not change;
-- the signed-drop approval stays unused, if that path was used;
-- manager counters return to their earlier values, if that path was used; and
-- no payment credit is recorded, if that path creates one.
-
-If the transaction succeeds, the token and the records for that path update together.
-
-### Missing ADR check
-
-The accepted design also requires valid artist permission for an artist-bound collection. The reviewed contracts do not yet enforce that check.
-
-**Why this matters:** A collector cannot receive a half-finished mint.`,
-  "publicReview.pages.artworkLifecycle.currentTokenIdentitySection": `## 7. The minted token gets a permanent ID
-
-### What happens
-
-When a mint succeeds, the Core gives the token two numbers:
-
-- a global ID across all Stream tokens; and
-- a serial number inside its collection.
-
-Both numbers stay with the token for its full history.
-
-### What happens if the token is burned
-
-Stream keeps two supply counts:
-
-- **Minted ever:** every token successfully minted, including burned tokens.
-- **Live supply:** tokens that have not been burned.
-
-Burning removes the token from current ownership and lowers the live supply.
-
-It does not lower the minted-ever count or make room for a replacement mint.
-
-The token’s ID, collection link, serial number, and burn record remain stored. Its ID is never reused.
-
-**Why this matters:** Burning a token does not erase its history or change the identity of other tokens.
-
-### Technical details
-
-The Core stores the collection ID and collection serial directly for each token. It does not calculate them from the token ID. This matches the accepted ADR design.`,
-  "publicReview.pages.artworkLifecycle.currentRemainingSections": `## 8. Randomness is requested and recorded
-
-### What happens
-
-If an artwork uses randomness, minting starts a request. The request is tied to the collection, token, provider, and provider version.
-
-The request can be:
-
-- **Waiting:** the provider has not answered yet.
-- **Complete:** the final value was accepted.
-- **Stale:** the old provider version can no longer finish it.
-- **Processing failed:** the provider answered, but the Core could not save the result.
-
-### What the current contracts record
-
-The contracts keep the request ID, times, final seed, a hash of the provider result, any failure hash, and the retry count.
-
-If saving the result fails, a retry uses the same saved seed. It does not ask for a new random value.
-
-**Why this matters:** A technical retry cannot become a hidden redraw.
-
-### Technical details
-
-The provider version is called an **epoch**. It lets Stream separate requests made before and after a provider change. A hash helps prove that later data matches the original result without storing all of that result onchain.
-
-## 9. Metadata describes the artwork
-
-### What happens
-
-Apps and marketplaces use \`tokenURI\` to read a token's artwork description.
-
-The current renderer can combine:
-
-- collection details;
-- scripts and dependencies;
-- token data, images, and traits; and
-- randomness, when the artwork uses it.
-
-### What can still be incomplete
-
-A \`tokenURI\` does not prove that every part is ready or preserved.
-
-- Randomness may still be waiting.
-- An image or other file may be stored elsewhere and become unavailable.
-- A script may need a browser or dependency that no longer works.
-- A hash can prove that retrieved bytes are correct, but it cannot keep those bytes available.
-
-The accepted ADR design adds clear pending and final metadata states. It also requires the final freeze record to identify every input needed to display the work.
-
-**Why this matters:** A useful artwork record must explain both what the token says and what is needed to display it.
-
-## 10. Sale money becomes balances to withdraw
-
-### What happens
-
-A sale records how much each recipient is owed. Recipients withdraw their money later.
-
-This means one recipient that cannot accept a payment does not block the sale for everyone else.
-
-### What the current contracts record
-
-The current fixed-price path records balances for the account that starts the sale, the protocol, and the curator reserve. The auction path records balances for the same groups and for bidder refunds.
-
-Every recorded balance must be backed by money held by the contract. Emergency withdrawal can use only extra money that is not owed to anyone.
-
-The repository also contains newer revenue modules for clearer split rules and supported assets. The current sale paths do not use those modules everywhere yet.
-
-**Why this matters:** A completed sale cannot lose track of who is owed money, and one failed withdrawal cannot block other people.
-
-### Technical details
-
-The ADRs call money owed by a contract a **liability**. A pull payment means the contract records that liability first and the recipient withdraws it in a separate transaction.
-
-## 11. An auction ends once
-
-### What happens
-
-When an auction starts, the auction contract holds the token. Each new highest bid replaces the previous leader. The previous bidder receives a balance they can withdraw.
-
-After the auction ends, anyone can trigger settlement.
-
-### Possible outcomes
-
-- **A winning bid:** the winner receives the token and the sale balances are recorded.
-- **No bids:** the token normally returns to the account that started the auction. If that account is another contract, it can use a separate claim path.
-- **Cancellation:** this is allowed only before the first bid and before the auction ends.
-
-Each outcome is terminal. This means the auction cannot settle, transfer, refund, or cancel a second time.
-
-**Why this matters:** The same token or payment cannot be handled twice.
-
-## 12. Burning affects more than ownership
-
-Section 7 explained that burning does not erase a token's ID. It also affects other parts of the artwork record.
-
-### What the current contracts do
-
-- The owner and live token are removed.
-- Live supply falls, but minted-ever supply does not.
-- The token ID, collection link, serial number, and burn record remain.
-- \`tokenURI\` is no longer available for the burned token.
-- A later valid randomness answer can be kept for audit, but it cannot bring the token back.
-- Existing sale, payment, and preservation history remains separate from ownership.
-
-The accepted burn-to-mint design also requires a burned token ID to be usable only once as proof for a later claim.
-
-**Why this matters:** Future readers can tell the difference between a token that never existed and one that was minted and later burned.
-
-## 13. Supply closes
-
-### What happens
-
-After minting ends and the required wait passes, an authorized Stream account can close supply.
-
-The current Core sets the maximum supply to the number minted so far. Burned tokens still count because supply closure uses minted-ever history. If nothing was minted, supply closes at zero.
-
-### What closes
-
-Every mint path must eventually ask the Core for the next collection serial. After supply closes, the Core has no serial left to give. It rejects any later attempt to create a token, whether the request comes through an old mint phase, signed approval, or auction registration.
-
-Supply closure does not freeze scripts, metadata, or other artwork details. That happens in later stages.
-
-### Difference from the ADR design
-
-The final supply is visible in current contract state. The accepted design also expects a clear closure event. The current \`setFinalSupply\` function does not emit its own supply-closed event.
-
-**Why this matters:** The collection's final edition size cannot grow later.
-
-## 14. The Core is permanently frozen
-
-### What happens
-
-\`freezeCollection\` first makes supply final and saves a hash of the frozen Core state. The Core then rejects changes inside that boundary.
-
-The freeze stops:
-
-- new mints and burns;
-- changes to the artist snapshot approval;
-- changes to the randomness module; and
-- covered collection and live-token metadata changes.
-
-### What can still happen
-
-Collectors can still transfer their tokens. New preservation evidence can also be added in its separate record system.
-
-Core freeze is not the same as full artwork finality. It locks the permanent Core boundary. Section 16 explains the later finality ceremony across the wider artwork record.
-
-**Why this matters:** People can see exactly which permanent data has stopped changing without assuming that every part of the artwork is already final.
-
-## 15. Preservation records can still be added
-
-### What happens
-
-Approved writers can add preservation records for files, storage locations, manifests, signatures, and other evidence.
-
-These records are append-only. A new record does not delete an older one. A separate pointer only makes the newest record easier to find.
-
-Preservation records stay open after the Core is frozen. This lets a future archive add recovery information without rewriting the artist's earlier record.
-
-### What a hash cannot do
-
-A hash can prove that a file matches an earlier commitment. It cannot keep the file online.
-
-The accepted ADR design therefore also expects the actual files in independent storage, plus enough runtime information to open or run the artwork. The current contract records evidence of preservation. It cannot by itself guarantee that an outside storage service stays available.
-
-**Why this matters:** Long-term preservation needs both proof and access to the real artwork materials.
-
-## 16. Artwork finality is the last ceremony
-
-### What happens
-
-A finality proposal records the exact artwork package to be made final. It also sets a waiting period and an expiry time.
-
-Before execution, the current registry checks that:
-
-- the collection status is closed;
-- burning is blocked;
-- the Core is frozen;
-- the exact proposed record still matches; and
-- the required content and the correct artist approval or platform record are present.
-
-Before the waiting period ends, a guardian can veto the proposal. An administrator can cancel it any time before it is executed or expires. If the execution window ends, the proposal expires.
-
-A cancelled or expired proposal can be scheduled again. A new artist approval is needed when the artist-signed finality record changes. The same unchanged record does not automatically need a new signature.
-
-After execution, the covered artwork state is final. Append-only preservation evidence can still be added without changing that final artwork package.
-
-**Why this matters:** Everyone gets time to inspect the exact irreversible action before it happens.
-
-## 17. Replaceable modules can have successors
-
-### What happens
-
-Some tools around the artwork may become old or stop working. This can include a renderer, storage route, or randomness provider.
-
-The accepted design replaces such a module with a new version. It does not silently edit the old contract. The old version remains readable, while the successor handles clearly assigned future work.
-
-A successor cannot rewrite the token's Core identity or any frozen artwork commitment.
-
-### What a safe change must explain
-
-- Which future actions move to the successor?
-- Which pending jobs and money stay with the old module?
-- Can both versions act at the same time?
-- How do signatures, counters, and replay protection work across the change?
-- Which old commitments remain binding?
-
-The current module system exposes a module type, version, schema hash, and link to the module it replaces. Signatures for an old contract do not automatically become valid in a new one.
-
-**Why this matters:** Stream can replace aging tools without changing the artwork's permanent identity or hiding its history.
-
-## What collectors should be able to see
-
-The product should show:
-
-- whether randomness is still waiting;
-- whether metadata is complete;
-- whether supply is closed;
-- whether the Core is frozen;
-- which preservation package applies;
-- whether artwork finality has happened; and
-- which replaceable module version is current.
-
-These are separate facts. For example, frozen does not also mean preserved or final.
-
-## Technical review checklist
-
-Reviewers should test what happens when:
-
-- a collection starts with the wrong artist, supply, or module;
-- signed details are changed, copied, expired, or reused;
-- a failed mint leaves a token, counter, approval, or payment changed;
-- mint paths together exceed the intended supply;
-- randomness is delayed, fails, or is redrawn during recovery;
-- metadata has a correct hash but missing or unusable files;
-- money owed by a contract is missing or cannot be withdrawn;
-- burning erases history or wrongly opens new supply;
-- supply closes but one mint path still works;
-- Core freeze leaves another change path open;
-- finality executes a different record from the reviewed proposal; or
-- a successor duplicates authority, pending work, money, or replay state.
-
-## Questions for reviewers
-
-1. Is the purpose of each stage clear?
-2. Which changes need artist approval?
-3. Which failures should undo the whole transaction, and which need a visible recovery state?
-4. Are supply closure, Core freeze, preservation, and artwork finality clearly separate?
-5. Which burn and history records must remain readable forever?
-6. What evidence must exist before finality can be scheduled?
-7. What must be proven before a successor becomes active?`,
-  "publicReview.pages.forArtists.title": "For Artists",
-  "publicReview.pages.forArtists.summary":
-    "For artists considering or preparing to publish their work through Stream.",
-  "publicReview.pages.rolesAndTrust.title": "Roles and Trust",
-  "publicReview.pages.rolesAndTrust.summary":
-    "Every role that can act, what it can change, and where trust remains.",
-  "publicReview.pages.whoCanDoWhat.title": "Who Can Do What",
-  "publicReview.pages.whoCanDoWhat.summary":
-    "Who can act, what each person or contract can change, and how those powers end.",
-  "publicReview.pages.curationAndTdhAuthorization.title":
-    "Curation and TDH Authorization",
-  "publicReview.pages.curationAndTdhAuthorization.summary":
-    "How offchain curation and TDH decisions become signed onchain authorization.",
-  "publicReview.pages.tokensCollectionsAndMinting.title":
-    "Tokens, Collections, and Minting",
-  "publicReview.pages.tokensCollectionsAndMinting.summary":
-    "The shared ERC-721 system for collections, token issuance, supply, and mint controls.",
-  "publicReview.pages.fixedPriceSalesAndAuctions.title":
-    "Fixed-Price Sales and Auctions",
-  "publicReview.pages.fixedPriceSalesAndAuctions.summary":
-    "The sale mechanisms, bidding rules, settlement paths, and edge cases.",
-  "publicReview.pages.revenueSplitsAndRoyalties.title":
-    "Revenue, Splits, and Royalties",
-  "publicReview.pages.revenueSplitsAndRoyalties.summary":
-    "Where primary-sale funds and secondary royalties go, and how recipients are configured.",
-  "publicReview.pages.randomness.title": "Randomness",
-  "publicReview.pages.randomness.summary":
-    "How unpredictable values enter the protocol and which outcomes depend on them.",
-  "publicReview.pages.metadataScriptsAndDependencies.title":
-    "Metadata, Scripts, and Dependencies",
-  "publicReview.pages.metadataScriptsAndDependencies.summary":
-    "How token presentation, generative scripts, and external dependencies are stored and referenced.",
-  "publicReview.pages.freezingPreservationAndArtworkFinality.title":
-    "Freezing, Preservation, and Artwork Finality",
-  "publicReview.pages.freezingPreservationAndArtworkFinality.summary":
-    "The mechanisms that move artwork data from editable to permanently fixed.",
-  "publicReview.pages.governancePausingAndSuccessors.title":
-    "Governance, Pausing, and Successors",
-  "publicReview.pages.governancePausingAndSuccessors.summary":
-    "How governance acts, emergencies are handled, and successor contracts are recognized.",
-  "publicReview.pages.changesEmergenciesAndFutureContracts.title":
-    "Changes, Emergencies, and Future Contracts",
-  "publicReview.pages.changesEmergenciesAndFutureContracts.summary":
-    "How Stream announces changes, stops problems, replaces service contracts, and ends powers permanently.",
-  "publicReview.pages.securityTestingAndKnownLimitations.title":
-    "Security, Testing, and Known Limitations",
-  "publicReview.pages.securityTestingAndKnownLimitations.summary":
-    "Current engineering evidence, unresolved findings, constraints, and pre-audit caveats.",
-  "publicReview.pages.currentImplementationAndReadiness.title":
-    "Current Implementation and Readiness",
-  "publicReview.pages.currentImplementationAndReadiness.summary":
-    "The authoritative record of what is connected, implemented, proposed, tested, audited, and still required before release.",
-  "publicReview.pages.whereDevelopmentStands.title": "Where Development Stands",
-  "publicReview.pages.whereDevelopmentStands.summary":
-    "What worked in this review snapshot, what was being connected, and the evidence required before launch.",
-  "publicReview.pages.communityReview.title": "Community Review",
-  "publicReview.pages.communityReview.summary":
-    "How to examine the proposal, frame actionable feedback, and follow the review record.",
-  "publicReview.community.editorial.heading":
-    "How to participate in the community review",
-  "publicReview.community.editorial.intro.purpose":
-    "Use this review to report an unclear promise, product problem, code bug, security risk, or missing proof before Stream's permanent rules are fixed.",
-  "publicReview.community.editorial.intro.noCode":
-    "You do not need Solidity knowledge. Explain what you expected, what you found, and why it matters.",
-  "publicReview.community.editorial.intro.source":
-    "This page covers review version `{reviewVersion}` and source commit {sourceLink}. Check both before commenting.",
-  "publicReview.community.editorial.people.heading": "Who should comment",
-  "publicReview.community.editorial.people.intro":
-    "Different experience reveals different problems:",
-  "publicReview.community.editorial.people.artists":
-    "Artists can review consent, creative control, preservation, recovery, and finality.",
-  "publicReview.community.editorial.people.collectors":
-    "Collectors can review ownership, artwork access, sale terms, and long-term promises.",
-  "publicReview.community.editorial.people.engineers":
-    "Contract, security, storage, browser, metadata, and infrastructure engineers can test technical and operational risks.",
-  "publicReview.community.editorial.people.product":
-    "Product, frontend, accessibility, and localization reviewers can find confusing or excluding workflows.",
-  "publicReview.community.editorial.people.community":
-    "Community members can ask whether Stream's powers and permanent choices match 6529's values.",
-  "publicReview.community.editorial.people.outro":
-    "Solidity knowledge is optional. An unclear promise or unfair workflow is useful feedback.",
-  "publicReview.community.editorial.scope.heading": "What to review",
-  "publicReview.community.editorial.scope.intro": "Useful feedback can cover:",
-  "publicReview.community.editorial.scope.artist":
-    "Artist identity, consent, collaborators, recovery, and estates.",
-  "publicReview.community.editorial.scope.sales":
-    "Collections, minting, sales, auctions, payments, splits, and royalties.",
-  "publicReview.community.editorial.scope.services":
-    "Randomness, metadata, storage, browsers, marketplaces, and indexers.",
-  "publicReview.community.editorial.scope.artwork":
-    "Artwork changes, preservation, collection freeze, and separate artwork finality.",
-  "publicReview.community.editorial.scope.governance":
-    "Roles, governance, emergency pauses, contract replacement, and powers that end permanently.",
-  "publicReview.community.editorial.scope.evidence":
-    "Threats, tests, launch setup, outside services, audit, and release evidence.",
-  "publicReview.community.editorial.scope.question":
-    "For every mechanism, ask what it protects, whether its power is limited, and whether a simpler design would work.",
-  "publicReview.community.editorial.safety.heading":
-    "Before reporting a security issue",
-  "publicReview.community.editorial.safety.policy":
-    "The current public-review policy allows possible Stream vulnerabilities to be reported in the public review Wave.",
-  "publicReview.community.editorial.safety.neverShare": "Never include:",
-  "publicReview.community.editorial.safety.credentials":
-    "Private keys, seed phrases, passwords, cookies, credentials, or API keys.",
-  "publicReview.community.editorial.safety.personal":
-    "Personal information that is not needed to explain the issue.",
-  "publicReview.community.editorial.safety.unrelated":
-    "Secrets or private data from another system.",
-  "publicReview.community.editorial.safety.attack":
-    "Instructions or data for attacking a live system outside this review.",
-  "publicReview.community.editorial.safety.otherProtocol":
-    "If the same technique affects another live protocol, keep the Stream report focused on Stream and coordinate the other disclosure separately.",
-  "publicReview.community.editorial.submit.heading": "How to submit feedback",
-  "publicReview.community.editorial.submit.intro":
-    "Start from the review page that best matches the issue:",
-  "publicReview.community.editorial.submit.page":
-    "Open the feedback panel on that page.",
-  "publicReview.community.editorial.submit.section":
-    "Choose the relevant section, or choose the whole page.",
-  "publicReview.community.editorial.submit.classify":
-    "Choose the closest feedback type and likely impact.",
-  "publicReview.community.editorial.submit.explain":
-    "Explain what happened and add evidence when you have it.",
-  "publicReview.community.editorial.submit.send":
-    "Submit the report. It becomes part of the public Stream review Wave.",
-  "publicReview.community.editorial.submit.crossModule":
-    "If the issue spans several contracts or pages, submit it from the main topic and link the other evidence.",
-  "publicReview.community.editorial.submit.short":
-    "Short reports are welcome. One precise question can reveal a missing rule.",
-  "publicReview.community.editorial.type.heading":
-    "Choose the closest feedback type",
-  "publicReview.community.editorial.type.intro":
-    "Choose the main reason for the report:",
-  "publicReview.community.editorial.type.question":
-    "Question — something is unclear or needs an official answer.",
-  "publicReview.community.editorial.type.documentation":
-    "Documentation — an explanation is wrong, incomplete, or hard to understand.",
-  "publicReview.community.editorial.type.artist":
-    "Artist workflow — the design does not fit a real artist's process.",
-  "publicReview.community.editorial.type.product":
-    "Product or UX — an action, result, or recovery path is confusing.",
-  "publicReview.community.editorial.type.protocol":
-    "Protocol design — a role, rule, lifecycle, or permanent choice should change.",
-  "publicReview.community.editorial.type.bug":
-    "Implementation bug — the code does not match the stated behavior.",
-  "publicReview.community.editorial.type.security":
-    "Possible security vulnerability — a path may allow theft, an unauthorized change, a permanent lock, manipulation, or takeover.",
-  "publicReview.community.editorial.type.testing":
-    "Testing or evidence gap — an important promise lacks a repeatable test or real-world proof.",
-  "publicReview.community.editorial.type.accessibility":
-    "Accessibility or localization — the review or product excludes a group of people.",
-  "publicReview.community.editorial.type.outro":
-    "Choose one main type. Use the report to explain where the issue crosses into other areas.",
-  "publicReview.community.editorial.impact.heading": "Choose the likely impact",
-  "publicReview.community.editorial.impact.intro":
-    "This is your first estimate, not a final decision:",
-  "publicReview.community.editorial.impact.critical":
-    "Critical — a credible risk of major asset loss, permanent unauthorized artwork change, or protocol-wide takeover.",
-  "publicReview.community.editorial.impact.high":
-    "High — a serious asset, authority, availability, or finality failure with a realistic path.",
-  "publicReview.community.editorial.impact.medium":
-    "Medium — a meaningful but narrower failure, or one with stronger conditions or a practical workaround.",
-  "publicReview.community.editorial.impact.low":
-    "Low — limited impact or an extra safety improvement.",
-  "publicReview.community.editorial.impact.informational":
-    "Informational — a clarification, convention, maintenance, or evidence improvement.",
-  "publicReview.community.editorial.impact.unknown":
-    "Pending assessment — use this when you are unsure about the impact.",
-  "publicReview.community.editorial.impact.outro":
-    "State what you know, what you assume, and what remains uncertain.",
-  "publicReview.community.editorial.report.heading": "Write a useful report",
-  "publicReview.community.editorial.report.intro":
-    "Answer as many of these questions as you can:",
-  "publicReview.community.editorial.report.expected": "What did you expect?",
-  "publicReview.community.editorial.report.observed":
-    "What did you find in the pinned code or review page?",
-  "publicReview.community.editorial.report.affected":
-    "Which person, collection, token, role, contract, field, or state is affected?",
-  "publicReview.community.editorial.report.preconditions":
-    "What must happen first, or what must an actor control?",
-  "publicReview.community.editorial.report.consequence":
-    "What could happen because of the issue?",
-  "publicReview.community.editorial.report.reproduction":
-    "Can you reproduce it with an example, call sequence, transaction, or test?",
-  "publicReview.community.editorial.report.fix":
-    "What change should be considered, and what remains uncertain?",
-  "publicReview.community.editorial.report.product":
-    "For a product or documentation issue, include the wording or screen state that caused the misunderstanding.",
-  "publicReview.community.editorial.report.artist":
-    "For an artist-workflow issue, describe the real studio, collaborator, signing, preservation, or estate process that is missing.",
-  "publicReview.community.editorial.technical.heading":
-    "Add technical evidence when needed",
-  "publicReview.community.editorial.technical.intro":
-    "This section is mainly for engineers and auditors. Use the generated Technical Reference when possible because it points to the pinned source.",
-  "publicReview.community.editorial.technical.reference":
-    "Link every important contract, function, or definition.",
-  "publicReview.community.editorial.technical.commit":
-    "If you use GitHub, link to the exact reviewed commit.",
-  "publicReview.community.editorial.technical.order":
-    "Describe the order of calls across contracts.",
-  "publicReview.community.editorial.technical.state":
-    "Explain which stored values change and which steps fail or remain.",
-  "publicReview.community.editorial.technical.test":
-    "Include the test, script, or settings needed to reproduce the issue.",
-  "publicReview.community.editorial.technical.paths":
-    "For a pause, replacement, successor, or finality bypass, include every other path that can create the same result.",
-  "publicReview.community.editorial.after.heading":
-    "What happens after submission",
-  "publicReview.community.editorial.after.new":
-    "Every new structured report appears in the review record with the state `NEW`.",
-  "publicReview.community.editorial.after.replies":
-    "Replies remain in the linked public Wave discussion.",
-  "publicReview.community.editorial.after.currentLimit":
-    "The current frontend does not yet record accepted, rejected, confirmed, or fixed states. Its only report state is `NEW`.",
-  "publicReview.community.editorial.after.future":
-    "A future response system should record who responded, the source change or decision, supporting tests, remaining risk, and the review version.",
-  "publicReview.community.editorial.record.heading":
-    "How the review record works",
-  "publicReview.community.editorial.record.intro":
-    "Each structured report stores four machine-readable fields:",
-  "publicReview.community.editorial.record.schema":
-    "`review_schema` — the format version used to read the report.",
-  "publicReview.community.editorial.record.type":
-    "`type` — the selected feedback type.",
-  "publicReview.community.editorial.record.severity":
-    "`severity` — the submitter's first impact estimate.",
-  "publicReview.community.editorial.record.context":
-    "`context` — the submission ID, review, version, page, optional section, and optional code reference.",
-  "publicReview.community.editorial.record.contextDetail":
-    "The code reference can pin a repository, commit, file, file hash, line range, contract, declaration, and selected-code hash.",
-  "publicReview.community.editorial.record.visible":
-    "The visible report uses normal human language. The stored fields support filtering and source checks.",
-  "publicReview.community.editorial.record.limit":
-    "The stored context does not include the frontend commit or the full wording shown on screen. By itself, it cannot prove the exact copy a person saw.",
-  "publicReview.community.editorial.record.filters":
-    "The current ledger can filter by feedback type, page, contract, impact, and the `NEW` state.",
-  "publicReview.community.editorial.versions.heading":
-    "When the source changes",
-  "publicReview.community.editorial.versions.new":
-    "A new contract candidate should receive a new immutable review bundle and version-specific routes.",
-  "publicReview.community.editorial.versions.old":
-    "Older versioned explanations, source links, and reports remain available.",
-  "publicReview.community.editorial.versions.manual":
-    "The current process compares review versions manually. It does not automatically decide whether an older report still applies.",
-  "publicReview.community.editorial.versions.carry":
-    "For each new candidate, record which earlier reports still apply, were fixed, or need another review.",
-  "publicReview.community.editorial.audit.heading":
-    "For auditors and review closeout",
-  "publicReview.community.editorial.audit.intro":
-    "The generated code inventory and public feedback record support an audit. They do not replace one.",
-  "publicReview.community.editorial.audit.currentHeading":
-    "What the current tools support",
-  "publicReview.community.editorial.audit.inventory":
-    "Reproduce the compiler-based code inventory.",
-  "publicReview.community.editorial.audit.reports":
-    "List reports that link to exact code.",
-  "publicReview.community.editorial.audit.filters":
-    "Filter reports by page, contract, type, impact, and `NEW` state.",
-  "publicReview.community.editorial.audit.export":
-    "Export the feedback records that have been loaded.",
-  "publicReview.community.editorial.audit.limit":
-    "Cross-version decisions and official response states are not automated. Auditors must compare them with published evidence.",
-  "publicReview.community.editorial.audit.closeoutHeading":
-    "What closeout should publish",
-  "publicReview.community.editorial.audit.candidate":
-    "The final candidate commit and Git tree.",
-  "publicReview.community.editorial.audit.versions":
-    "Every review version and available structural comparison.",
-  "publicReview.community.editorial.audit.risks":
-    "Unresolved questions and accepted risks.",
-  "publicReview.community.editorial.audit.fixes":
-    "Fix commits and repeatable test evidence.",
-  "publicReview.community.editorial.audit.auditReport":
-    "Audit reports and remediation status.",
-  "publicReview.community.editorial.audit.blockers":
-    "The final release-blocker checklist.",
-  "publicReview.community.editorial.audit.archive":
-    "A machine-readable archive of the review record.",
-  "publicReview.community.editorial.audit.outro":
-    "Independent experts must still audit the exact release candidate and launch configuration.",
-  "publicReview.development.heading": "Is Stream ready to launch?",
-  "publicReview.development.answer": "Not yet.",
-  "publicReview.development.summary":
-    "Stream has working contracts and many tests. But important safety checks are still missing.",
-  "publicReview.development.beforeLaunch":
-    "Before launch, Stream still needs:",
-  "publicReview.development.beforeLaunch.audit":
-    "An independent security audit.",
-  "publicReview.development.beforeLaunch.liveTesting":
-    "Testing with real services and marketplaces.",
-  "publicReview.development.beforeLaunch.launchSetup":
-    "A final, verified launch setup.",
-  "publicReview.development.beforeLaunch.blockers":
-    "All serious release blockers resolved.",
-  "publicReview.development.pagePurpose":
-    "This page explains what works, what is still uncertain, and what must happen before Stream can launch.",
-  "publicReview.development.lastChecked": "Last checked",
-  "publicReview.development.openBlockers": "Open release blockers",
-  "publicReview.development.editorial.scopeHeading":
-    "What this review covers",
-  "publicReview.development.editorial.scopeSummary":
-    "Stream was not live when this review was created, and no Stream contracts held funds.",
-  "publicReview.development.editorial.progress.heading":
-    "How we describe progress",
-  "publicReview.development.editorial.progress.intro":
-    "We track two separate things:",
-  "publicReview.development.editorial.progress.built":
-    "What has been built.",
-  "publicReview.development.editorial.progress.proof":
-    "What proof exists that it is safe and ready.",
-  "publicReview.development.editorial.progress.caveat":
-    "A progress label shows how far the work has moved. It does not mean Stream is ready to launch.",
-  "publicReview.development.editorial.progress.labelHeading": "Label",
-  "publicReview.development.editorial.progress.meaningHeading":
-    "What it means",
-  "publicReview.development.editorial.progress.working.label":
-    "Working in the reviewed flow",
-  "publicReview.development.editorial.progress.working.meaning":
-    "Built and connected in the full user path described here.",
-  "publicReview.development.editorial.progress.connected.label":
-    "Connected to selected parts",
-  "publicReview.development.editorial.progress.connected.meaning":
-    "Built and connected to some Stream parts. The full user path may still be incomplete.",
-  "publicReview.development.editorial.progress.code.label": "Code exists",
-  "publicReview.development.editorial.progress.code.meaning":
-    "The code exists in the reviewed version. Launch setup and safety proof may still be missing.",
-  "publicReview.development.editorial.progress.plan.label": "Accepted plan",
-  "publicReview.development.editorial.progress.plan.meaning":
-    "The design direction is accepted. Complete code or connection proof is still missing.",
-  "publicReview.development.editorial.progress.open.label": "Not final",
-  "publicReview.development.editorial.progress.open.meaning":
-    "The design is still open, or the work belongs to a later version.",
-  "publicReview.development.editorial.working.heading":
-    "What works in the reviewed flow",
-  "publicReview.development.editorial.working.intro":
-    "These paths are connected in the exact reviewed setup. This is a rehearsal, not a live launch.",
-  "publicReview.development.editorial.working.identity.heading":
-    "Artwork identity and history",
-  "publicReview.development.editorial.working.identity.summary":
-    "Core keeps each token's identity, collection, supply, artist approval, metadata state, and burn history. It also records collection freezes, but a freeze is not complete artwork finality.",
-  "publicReview.development.editorial.working.sales.heading":
-    "Fixed-price sales and auctions",
-  "publicReview.development.editorial.working.sales.summary":
-    "A fixed-price sale checks a signed approval and uses the existing minting contract. An auction creates the token inside the auction contract before bidding starts.",
-  "publicReview.development.editorial.working.payments.heading":
-    "Payments and royalties",
-  "publicReview.development.editorial.working.payments.summary":
-    "Fixed-price sales record credits for the seller, protocol, and curator reserve. Auctions track bidder refunds. Core reports one royalty receiver and a 6.9% royalty for every token; the newer revenue resolver, which stores revenue rules, is not used for this check.",
-  "publicReview.development.editorial.working.safety.heading":
-    "Safety controls and preservation",
-  "publicReview.development.editorial.working.safety.summary":
-    "The admin contract can pause everything or selected actions. Separate roles control emergency pauses and unpausing. Core and the preservation-record contract are included, but artwork recovery and every launch role are not yet proven.",
-  "publicReview.development.editorial.working.boundary":
-    "Only connected behavior is counted here. ADR proposals are not working features.",
-  "publicReview.development.editorial.working.setupLink":
-    "See the exact rehearsal setup",
-  "publicReview.development.editorial.connected.heading":
-    "Connected to selected parts",
-  "publicReview.development.editorial.connected.intro":
-    "Some newer systems are connected to selected contracts. The main Drop and Auction flow does not use them yet.",
-  "publicReview.development.editorial.connected.minting.heading":
-    "A newer minting path",
-  "publicReview.development.editorial.connected.minting.summary":
-    "The mint manager and ledger control who may mint, when minting is open, and supply limits. The ledger keeps past-mint counters. The manager is connected to Core and can update the ledger.",
-  "publicReview.development.editorial.connected.minting.missing":
-    "Signed Drops still use the existing minting contract. The old and new minting paths remain separate.",
-  "publicReview.development.editorial.connected.minting.link":
-    "Learn how the two minting paths differ",
-  "publicReview.development.editorial.connected.payments.heading":
-    "A newer payment path",
-  "publicReview.development.editorial.connected.payments.summary":
-    "The reviewed setup includes newer contracts for payment rules, payment splits, allowed assets, and sale settlement.",
-  "publicReview.development.editorial.connected.payments.missing":
-    "Drops and Auctions still keep their own payment records, and settlement has no configured caller. Sale credits, payment splits, and Core's royalty information remain separate paths.",
-  "publicReview.development.editorial.connected.payments.link":
-    "Learn how the payment paths differ",
-  "publicReview.development.editorial.connected.boundary":
-    "These connections do not prove a complete user flow. ADR proposals are not counted here.",
-  "publicReview.development.editorial.code.heading": "Code exists",
-  "publicReview.development.editorial.code.intro":
-    "The reviewed source contains these features. The exact launch setup has not been proven to use them.",
-  "publicReview.development.editorial.code.existsLabel": "What exists:",
-  "publicReview.development.editorial.code.missingLabel":
-    "What is missing:",
-  "publicReview.development.editorial.code.governance.heading":
-    "Governance and contract replacements",
-  "publicReview.development.editorial.code.governance.exists":
-    "Governance code can delay, cancel, expire, or block actions. It can permanently freeze selected functions and close the initial setup stage. A registry can record each contract's identity, code, status, and approved replacement.",
-  "publicReview.development.editorial.code.governance.missing":
-    "The launch setup has not been proven to include the complete contract list, roles, delays, settings, and replacement paths.",
-  "publicReview.development.editorial.code.artwork.heading":
-    "Artwork protection and records",
-  "publicReview.development.editorial.code.artwork.exists":
-    "Artwork code can delay a permanent final state and allow cancellation or guardian blocking before it becomes final. Other code controls artwork and preservation writers and can create permanent metadata snapshots.",
-  "publicReview.development.editorial.code.artwork.missing":
-    "The launch setup has not been proven to use these systems. Final artwork data, writers, permissions, live providers, and recovery are not fully defined or independently reviewed.",
-  "publicReview.development.editorial.code.minting.heading":
-    "Minting rules",
-  "publicReview.development.editorial.code.minting.exists":
-    "Some governed settings can only increase after a delay, by no more than double, with no emergency shortcut. The minting system can check permissions, limits, and counters, then prepare and complete a mint as one operation.",
-  "publicReview.development.editorial.code.minting.missing":
-    "The full settings list is not bound to the launch setup. Signed Drops still use the older minting path, and some planned gate protections are not implemented.",
-  "publicReview.development.editorial.code.randomness.heading":
-    "Randomness and metadata",
-  "publicReview.development.editorial.code.randomness.exists":
-    "Randomness code tracks waiting, completed, stale, and failed requests and reuses the same random result for a retry. Metadata code supports data stored on the blockchain or outside it, plus scripts, images, attributes, and update notices.",
-  "publicReview.development.editorial.code.randomness.missing":
-    "Live randomness funding, callbacks, monitoring, stale-request rules, and recovery are not proven. Metadata still needs real-world proof with public services, browsers, marketplaces, large responses, and long-term storage.",
-  "publicReview.development.editorial.code.boundary":
-    "This is available code, not active launch protection. ADR proposals are not counted as working features.",
-  "publicReview.development.editorial.plan.heading": "Accepted plan",
-  "publicReview.development.editorial.plan.intro":
-    "The direction is accepted, but complete code and launch proof are still missing.",
-  "publicReview.development.editorial.plan.plannedLabel":
-    "What is planned:",
-  "publicReview.development.editorial.plan.whyLabel": "Why it helps:",
-  "publicReview.development.editorial.plan.missingLabel":
-    "What is still missing:",
-  "publicReview.development.editorial.plan.revenue.heading":
-    "Safer revenue checks",
-  "publicReview.development.editorial.plan.revenue.planned":
-    "An immutable validation contract would check revenue changes. It would hold no funds, roles, or Stream data and could read only from approved contracts with fixed code.",
-  "publicReview.development.editorial.plan.revenue.why":
-    "The revenue resolver would stay in control and reject failed checks, changed code, or bad answers before saving anything.",
-  "publicReview.development.editorial.plan.revenue.missing":
-    "The final validation contract and matching resolver code do not exist in this reviewed version. Their interface and code need independent approval. Recovery would also require a new resolver, continuity proof, registry approval, and a governed Core update.",
-  "publicReview.development.editorial.plan.metadata.heading":
-    "Metadata refresh from approved contracts",
-  "publicReview.development.editorial.plan.metadata.planned":
-    "Approved Stream contracts will be able to send standard metadata refresh notices for one token or a group of tokens.",
-  "publicReview.development.editorial.plan.metadata.reason":
-    "Marketplaces and other services can learn that token metadata changed and should be loaded again.",
-  "publicReview.development.editorial.plan.metadata.missing":
-    "Core currently sends these notices only for its own changes. The new public helpers do not exist. Permissions, lifecycle rules, token limits, abuse protection, implementation, and tests are still required.",
-  "publicReview.development.editorial.plan.metadata.link":
-    "Learn more about metadata refresh",
-  "publicReview.development.editorial.plan.roles.heading":
-    "Exact launch roles",
-  "publicReview.development.editorial.plan.roles.planned":
-    "The launch design lists 37 roles. Each must be connected to exact accounts or contracts, allowed actions, scopes, delays, and removal rules.",
-  "publicReview.development.editorial.plan.roles.reason":
-    "The launch team and reviewers can see exactly who can do what.",
-  "publicReview.development.editorial.plan.roles.missing":
-    "The final launch candidate has not yet proven all 37 role assignments and limits.",
-  "publicReview.development.editorial.plan.boundary":
-    "\"Accepted plan\" means the direction is agreed. It does not mean the work is built, connected, tested, or ready for launch.",
-  "publicReview.development.editorial.open.heading": "Not final",
-  "publicReview.development.editorial.open.intro":
-    "These are proposals or later-version ideas. They are not launch features.",
-  "publicReview.development.editorial.open.proposalLabel":
-    "What is proposed:",
-  "publicReview.development.editorial.open.missingLabel":
-    "What is missing:",
-  "publicReview.development.editorial.open.artist.heading":
-    "Artist permissions and recovery",
-  "publicReview.development.editorial.open.artist.proposal":
-    "ADR 0022 proposes an immutable, stateless validation adapter. It would not be registered, store data, or hold authority. The artist registry would remain the only authority and state owner.",
-  "publicReview.development.editorial.open.artist.missing":
-    "The adapter is not implemented and needs separate approval. Broader rules for collaborators, delegates, guardians, estates, sanctions, disputes, and account recovery are also not final.",
-  "publicReview.development.editorial.open.payments.heading":
-    "Payments and sale types",
-  "publicReview.development.editorial.open.payments.proposal":
-    "A proposed ERC-20 token sale contract would check the payer's signed payment instruction and be the only Stream contract allowed to pull the payment. Settlement would only record and distribute it.",
-  "publicReview.development.editorial.open.payments.missing":
-    "The verifier and complete sale path are not implemented. Dutch auctions, private sales, refund windows, sealed bids, raffles, burn-to-mint, and ERC-20 bidding are proposed or deferred, not current candidate features.",
-  "publicReview.development.editorial.open.records.heading":
-    "Mint records and artwork recovery",
-  "publicReview.development.editorial.open.records.proposal":
-    "ADR 0018 would make the ledger the permanent owner of each batch replay record and connect the ledger record to each token's prepared-mint event. ADR 0020 would preserve the original artwork-finality record and add governed recovery history without replacing or hiding the original.",
-  "publicReview.development.editorial.open.records.missing":
-    "Both proposals still need acceptance and implementation.",
-  "publicReview.development.editorial.open.randomness.heading":
-    "Randomness recovery and providers",
-  "publicReview.development.editorial.open.randomness.proposal":
-    "One proposal would add a fixed waiting period and one limited recovery step for stale randomness requests.",
-  "publicReview.development.editorial.open.randomness.missing":
-    "In the current code, a stale request is final for that token. RandomizerNXT code exists at the reviewed commit, but it is not an approved production provider in the reviewed release policy.",
-  "publicReview.development.editorial.open.boundary":
-    "None of these proposals are protections provided by the launch candidate.",
-  "publicReview.development.editorial.remaining.tests.heading":
-    "Test evidence",
-  "publicReview.development.editorial.remaining.tests.intro":
-    "The pinned code has several kinds of local tests:",
-  "publicReview.development.editorial.remaining.tests.unit":
-    "Unit tests check one behavior at a time.",
-  "publicReview.development.editorial.remaining.tests.fuzz":
-    "Fuzz and invariant tests try many inputs and check rules that should always hold.",
-  "publicReview.development.editorial.remaining.tests.composition":
-    "State-machine and composition tests combine actions and contracts.",
-  "publicReview.development.editorial.remaining.tests.focused":
-    "Focused tests cover signatures, auctions, mint accounting, burns, metadata, randomness, governance, and preservation.",
-  "publicReview.development.editorial.remaining.tests.provesHeading":
-    "What this proves",
-  "publicReview.development.editorial.remaining.tests.proves":
-    "The pinned code has meaningful local test evidence. The Technical Reference lists the exact compiled code and declarations.",
-  "publicReview.development.editorial.remaining.tests.limitsHeading":
-    "What this does not prove",
-  "publicReview.development.editorial.remaining.tests.limits":
-    "Tests can pass while an important risk remains. Reviewers still need to look for:",
-  "publicReview.development.editorial.remaining.tests.limit.assertions":
-    "Missing checks and untested combinations.",
-  "publicReview.development.editorial.remaining.tests.limit.specification":
-    "A wrong rule implemented the same way in both code and tests.",
-  "publicReview.development.editorial.remaining.tests.limit.deployment":
-    "Deployment mistakes, economic attacks, and lost or compromised keys.",
-  "publicReview.development.editorial.remaining.tests.limit.external":
-    "Failures in real providers, storage, browsers, public blockchain services, or marketplaces.",
-  "publicReview.development.editorial.remaining.tests.gapsHeading":
-    "Known test gaps",
-  "publicReview.development.editorial.remaining.tests.mintGap":
-    "The old and new minting paths are tested separately. Connecting a signed Drop to the new mint manager needs its own integration tests.",
-  "publicReview.development.editorial.remaining.tests.revenueGap":
-    "Revenue components are also tested separately. One joined settlement path needs its own integration tests.",
-  "publicReview.development.editorial.remaining.static.heading":
-    "Static analysis",
-  "publicReview.development.editorial.remaining.static.definition":
-    "Static analysis scans code for risky patterns without running it.",
-  "publicReview.development.editorial.remaining.static.count":
-    "The pinned Slither report has 30 open High or Medium findings: 3 High and 27 Medium.",
-  "publicReview.development.editorial.remaining.static.caveat":
-    "Some tool findings may be false positives. Even so, the candidate cannot be called clean until every finding has one clear outcome:",
-  "publicReview.development.editorial.remaining.static.fix":
-    "Fix it and add a test.",
-  "publicReview.development.editorial.remaining.static.falsePositive":
-    "Prove that it is a false positive.",
-  "publicReview.development.editorial.remaining.static.acceptedRisk":
-    "Accept the risk and record its owner and scope.",
-  "publicReview.development.editorial.remaining.static.remove":
-    "Remove the affected code from the release.",
-  "publicReview.development.editorial.remaining.static.register":
-    "The final record must keep the tool version, settings, raw output, source commit, and response evidence.",
-  "publicReview.development.editorial.remaining.static.link":
-    "Open the pinned Slither report",
-  "publicReview.development.editorial.remaining.limitations.heading":
-    "Known limitations",
-  "publicReview.development.editorial.remaining.limitations.intro":
-    "These are current code or evidence gaps. Proposed ADRs are not counted as fixes.",
-  "publicReview.development.editorial.remaining.limitations.minting.heading":
-    "Tokens and minting",
-  "publicReview.development.editorial.remaining.limitations.minting.finalSupply":
-    "A final supply of zero is also treated as not set. Before a collection is frozen, a later update can set a nonzero cap and reopen minting. There is no separate final-supply flag or event.",
-  "publicReview.development.editorial.remaining.limitations.minting.paths":
-    "Signed Drops and the current Auction use the older minter. The manager and ledger are a separate minting path.",
-  "publicReview.development.editorial.remaining.limitations.minting.nullifiers":
-    "The gate interface supports one-use nullifiers, but the current manager accepts empty lists. Nullifier-backed gates are not implemented.",
-  "publicReview.development.editorial.remaining.limitations.minting.replay":
-    "Mint replay data is split between the ledger and Core. ADR 0018 proposes one durable joined record, but it is not accepted or implemented.",
-  "publicReview.development.editorial.remaining.limitations.minting.identity":
-    "All collections share one Core token contract. ADR 0016 accepts this Core-only design; there is no hidden contract for each collection. Marketplace and indexer support still needs real-world proof.",
-  "publicReview.development.editorial.remaining.limitations.minting.link":
-    "Read the full token and minting explanation",
-  "publicReview.development.editorial.remaining.limitations.payments.heading":
-    "Payments and auctions",
-  "publicReview.development.editorial.remaining.limitations.payments.paths":
-    "Drops, Auctions, the curator pool, resolver, split wallets, settlement, and Core royalties still use separate money paths and records.",
-  "publicReview.development.editorial.remaining.limitations.payments.asset":
-    "An approved settlement caller can choose any active payment token. The payment token is not part of the sale's replay key, so the first successful settlement uses that shared key.",
-  "publicReview.development.editorial.remaining.limitations.payments.auction":
-    "The minimum bid increase and auction extension time are global settings. They have no bound, delay, or change event and can change during an active auction.",
-  "publicReview.development.editorial.remaining.limitations.payments.link":
-    "Read the full revenue explanation",
-  "publicReview.development.editorial.remaining.limitations.governance.heading":
-    "Governance and contract size",
-  "publicReview.development.editorial.remaining.limitations.governance.records":
-    "The candidate still lacks complete proof for admitted record types, authority providers, grants, live bindings, rotation, and revocation.",
-  "publicReview.development.editorial.remaining.limitations.governance.value":
-    "The governance executor can send native ETH too broadly, or its limits are not proven clearly enough.",
-  "publicReview.development.editorial.remaining.limitations.governance.binding":
-    "There is not yet end-to-end proof that every governed action binds every sensitive value.",
-  "publicReview.development.editorial.remaining.limitations.governance.size":
-    "StreamCore has only 424 bytes below the chain's runtime-code limit. The detailed size section explains the remaining margin risk.",
-  "publicReview.development.editorial.remaining.limitations.governance.link":
-    "Read the full governance explanation",
-  "publicReview.development.editorial.remaining.limitations.randomness.heading":
-    "Randomness",
-  "publicReview.development.editorial.remaining.limitations.randomness.stale":
-    "An authorized admin can mark a pending request stale immediately. Stale is final for that token.",
-  "publicReview.development.editorial.remaining.limitations.randomness.retry":
-    "A failed post-processing step can reduce the pending count before retry. Replacing the provider can then prevent the old request from passing retry checks.",
-  "publicReview.development.editorial.remaining.limitations.randomness.target":
-    "Provider adapters can change their Core target. A request keeps its provider and epoch, but not the Core address used when the request was assigned.",
-  "publicReview.development.editorial.remaining.limitations.randomness.live":
-    "Funding, callbacks, permissions, monitoring, stale rules, and raw-word recovery still need live proof. RandomizerNXT is not in the approved production provider set.",
-  "publicReview.development.editorial.remaining.limitations.randomness.link":
-    "Read the full randomness explanation",
-  "publicReview.development.editorial.remaining.limitations.artwork.heading":
-    "Artwork and metadata",
-  "publicReview.development.editorial.remaining.limitations.artwork.refresh":
-    "Core sends refresh notices for its own changes. The accepted helper for approved satellite contracts is not implemented. Collection-wide refresh also covers extra token IDs when collections are interleaved.",
-  "publicReview.development.editorial.remaining.limitations.artwork.size":
-    "Large metadata works in local tests, but public services, wallets, indexers, and marketplaces still need maximum-size tests.",
-  "publicReview.development.editorial.remaining.limitations.artwork.finality":
-    "Artwork finality can be scheduled or vetoed. The exact final data and every contract allowed to write it are not yet proven. Collection freeze is a separate protection.",
-  "publicReview.development.editorial.remaining.limitations.artwork.availability":
-    "A content hash proves that retrieved bytes are correct. It does not prove that the bytes, dependencies, browser, blockchain service, or rendering tools will remain available.",
-  "publicReview.development.editorial.remaining.limitations.artwork.link":
-    "Read the full artwork and preservation explanation",
-  "publicReview.development.editorial.remaining.standard.heading":
-    "The release standard",
-  "publicReview.development.editorial.remaining.standard.identity":
-    "Stream must keep token identity and artistic intent safe while replaceable sales, randomness, metadata, and operating services change.",
-  "publicReview.development.editorial.remaining.standard.evidence":
-    "Every mechanism needs one clear purpose, limited powers, and evidence that an independent person can check. Release depends on that evidence, not on labels or plans.",
-  "publicReview.development.editorial.remaining.bytecode.heading":
-    "Contract bytecode size",
-  "publicReview.development.editorial.remaining.bytecode.coreHeading":
-    "StreamCore now",
-  "publicReview.development.editorial.remaining.bytecode.core":
-    "The pinned build measures StreamCore at 24,152 bytes. The chain limit is 24,576 bytes, so 424 bytes remain. The interim target keeps 384 bytes free, leaving only 40 extra bytes. The long-term 2,000-byte target is still 1,576 bytes away.",
-  "publicReview.development.editorial.remaining.bytecode.risk":
-    "Small code, compiler, optimizer, or metadata changes could cross the limit. A small margin also makes emergency fixes harder.",
-  "publicReview.development.editorial.remaining.bytecode.futureHeading":
-    "Accepted future revenue design",
-  "publicReview.development.editorial.remaining.bytecode.future":
-    "ADR 0021 limits both the future resolver and validation adapter to 22,576 bytes of runtime code and 47,152 bytes of full deployment code, including constructor values.",
-  "publicReview.development.editorial.remaining.bytecode.notBuilt":
-    "Neither contract exists at the pinned commit. Only a final isolated build can prove its size.",
-  "publicReview.development.editorial.remaining.bytecode.link":
-    "Open the pinned bytecode proof",
-  "publicReview.development.editorial.remaining.candidate.heading":
-    "Evidence required for the exact release candidate",
-  "publicReview.development.editorial.remaining.candidate.intro":
-    "The final candidate needs a machine-readable manifest. It must include:",
-  "publicReview.development.editorial.remaining.candidate.codeHeading":
-    "Code identity",
-  "publicReview.development.editorial.remaining.candidate.code.hashes":
-    "Exact source and artifact hashes, plus compiler and optimizer settings.",
-  "publicReview.development.editorial.remaining.candidate.code.deployment":
-    "Chain, contract addresses, creation transactions, runtime code hashes, and constructor or initializer values.",
-  "publicReview.development.editorial.remaining.candidate.code.graph":
-    "Core pointers and the complete contract connection map.",
-  "publicReview.development.editorial.remaining.candidate.authorityHeading":
-    "Authority",
-  "publicReview.development.editorial.remaining.candidate.authority.roles":
-    "Role holders, scopes, signers, signer versions, pause controls, and guardians.",
-  "publicReview.development.editorial.remaining.candidate.authority.records":
-    "Admitted record types, authority providers, and grants.",
-  "publicReview.development.editorial.remaining.candidate.authority.governance":
-    "Governed settings and every allowed action type.",
-  "publicReview.development.editorial.remaining.candidate.externalHeading":
-    "External services and money",
-  "publicReview.development.editorial.remaining.candidate.external.randomness":
-    "Randomness providers, accounts, funding, callbacks, and monitoring.",
-  "publicReview.development.editorial.remaining.candidate.external.money":
-    "Revenue, settlement, splits, royalties, and every liability path.",
-  "publicReview.development.editorial.remaining.candidate.deploymentHeading":
-    "Deployment proof",
-  "publicReview.development.editorial.remaining.candidate.deployment.sealing":
-    "Ownership transfers, renounced powers, one-way setup sealing, and explorer verification.",
-  "publicReview.development.editorial.remaining.candidate.deployment.readback":
-    "An independent readback of every critical rule and connection.",
-  "publicReview.development.editorial.remaining.candidate.independent":
-    "A second person must be able to rebuild and verify the setup from the scripts and published evidence.",
-  "publicReview.development.editorial.remaining.external.heading":
-    "Evidence still required from real services",
-  "publicReview.development.editorial.remaining.external.intro":
-    "Local mocks are controlled substitutes. They do not prove that real services behave the same way.",
-  "publicReview.development.editorial.remaining.external.chainHeading":
-    "Chain and wallets",
-  "publicReview.development.editorial.remaining.external.chain.testnet":
-    "Run the candidate on a fork or public test network with the intended infrastructure.",
-  "publicReview.development.editorial.remaining.external.chain.wallets":
-    "Test production-style signing and contract wallets.",
-  "publicReview.development.editorial.remaining.external.servicesHeading":
-    "Providers and public services",
-  "publicReview.development.editorial.remaining.external.services.randomness":
-    "Test live randomness requests, callbacks, provider changes, and revoked access.",
-  "publicReview.development.editorial.remaining.external.services.marketplaces":
-    "Test shared collection identity, metadata, and royalties in marketplaces and indexers.",
-  "publicReview.development.editorial.remaining.external.services.rpc":
-    "Test the largest token metadata response through public blockchain services.",
-  "publicReview.development.editorial.remaining.external.operationsHeading":
-    "Long-running operations",
-  "publicReview.development.editorial.remaining.external.operations.preservation":
-    "Recover preservation packages independently from published instructions.",
-  "publicReview.development.editorial.remaining.external.operations.auctions":
-    "Run long auctions and test failure recovery.",
-  "publicReview.development.editorial.remaining.external.operations.continuity":
-    "Test signer changes, authority-provider changes, successor discovery, and continuity.",
-  "publicReview.development.editorial.remaining.external.coverage":
-    "These tests must cover real limits, permissions, delays, billing, availability, and failures.",
-  "publicReview.development.editorial.remaining.audit.heading":
-    "Independent audit",
-  "publicReview.development.editorial.remaining.audit.status":
-    "The exact release candidate still needs an independent audit and a completed record of every fix or accepted risk. Community review helps, but it does not replace the audit.",
-  "publicReview.development.editorial.remaining.audit.scope":
-    "The audit must cover the exact code and launch setup. Any important change after the audit needs a focused follow-up review.",
-  "publicReview.development.editorial.remaining.audit.record":
-    "Each finding needs a source-backed response, a fix commit where needed, a regression test, and a record of any risk that remains.",
-  "publicReview.development.editorial.remaining.blockers.heading":
-    "Release blockers",
-  "publicReview.development.editorial.remaining.blockers.intro":
-    "Production release stays blocked until the following work has evidence:",
-  "publicReview.development.editorial.remaining.blockers.codeHeading":
-    "Fix known code risks",
-  "publicReview.development.editorial.remaining.blockers.code.findings":
-    "Resolve every High finding and every Medium finding that can materially affect the release.",
-  "publicReview.development.editorial.remaining.blockers.code.finalSupply":
-    "Fix the zero-mint final-supply behavior or redefine the promise accurately.",
-  "publicReview.development.editorial.remaining.blockers.code.governance":
-    "Prove record-type authority, action binding, and safe native ETH limits.",
-  "publicReview.development.editorial.remaining.blockers.code.bytecode":
-    "Meet the chosen Core bytecode margin or deliberately change the design.",
-  "publicReview.development.editorial.remaining.blockers.pathHeading":
-    "Choose one launch path",
-  "publicReview.development.editorial.remaining.blockers.path.minting":
-    "Choose the signed-sale minting path and test its callers, counters, replay data, and Core entry end to end.",
-  "publicReview.development.editorial.remaining.blockers.path.accounting":
-    "Keep or replace the Drop and Auction accounting paths deliberately. Do not leave two unclear routes for the same value.",
-  "publicReview.development.editorial.remaining.blockers.path.royalties":
-    "Keep or replace the current fixed royalties deliberately, then verify marketplace behavior.",
-  "publicReview.development.editorial.remaining.blockers.path.adapter":
-    "For ADR 0021, approve the final interface, build matching resolver and adapter code, review both, prove their sizes, and prove adapter-first deployment.",
-  "publicReview.development.editorial.remaining.blockers.setupHeading":
-    "Prove the exact setup",
-  "publicReview.development.editorial.remaining.blockers.setup.randomness":
-    "Test every production randomness provider outside local mocks and approve its stale and failure rules.",
-  "publicReview.development.editorial.remaining.blockers.setup.finality":
-    "Prove the exact artwork-finality data and every contract that can write a final state.",
-  "publicReview.development.editorial.remaining.blockers.setup.deployment":
-    "Rehearse deployment, setup sealing, rollback, and successor changeover, then read them back independently.",
-  "publicReview.development.editorial.remaining.blockers.setup.preservation":
-    "Recover preservation packages using only published instructions and commitments.",
-  "publicReview.development.editorial.remaining.blockers.setup.feedback":
-    "If production feedback is activated, bind and independently verify its one approved Wave.",
-  "publicReview.development.editorial.remaining.blockers.reviewHeading":
-    "Finish independent review",
-  "publicReview.development.editorial.remaining.blockers.review.audit":
-    "Complete the external audit and remediation record.",
-  "publicReview.development.editorial.remaining.blockers.review.version":
-    "Make the final public review match the exact code and configuration deployed.",
-  "publicReview.development.editorial.remaining.blockers.outro":
-    "Public review can add blockers. A blocker should be removed only when evidence shows that it is resolved.",
-  "publicReview.development.editorial.remaining.threats.heading":
-    "Threat model",
-  "publicReview.development.editorial.remaining.threats.intro":
-    "Assume that users, privileged accounts, external services, and ordinary people can act badly or fail.",
-  "publicReview.development.editorial.remaining.threats.attackersHeading":
-    "Malicious actions",
-  "publicReview.development.editorial.remaining.threats.attackers.people":
-    "Malicious buyers, bidders, recipients, and contract wallets.",
-  "publicReview.development.editorial.remaining.threats.attackers.chain":
-    "Front-running, chain reorganizations, changing timestamps, and transaction censorship.",
-  "publicReview.development.editorial.remaining.threats.attackers.signatures":
-    "Old, replayed, or partly bound signatures, and governance text that does not match the executed bytes.",
-  "publicReview.development.editorial.remaining.threats.attackers.money":
-    "Reentrancy, forced value, and recipients that reject ETH or tokens.",
-  "publicReview.development.editorial.remaining.threats.systemsHeading":
-    "Compromised or failing systems",
-  "publicReview.development.editorial.remaining.threats.systems.accounts":
-    "Compromised or mistaken privileged accounts.",
-  "publicReview.development.editorial.remaining.threats.systems.services":
-    "Unavailable or hostile randomness, storage, blockchain services, browsers, marketplaces, and indexers.",
-  "publicReview.development.editorial.remaining.threats.systems.metadata":
-    "Malformed or very large metadata.",
-  "publicReview.development.editorial.remaining.threats.systems.config":
-    "Registries or contracts connected in the wrong order.",
-  "publicReview.development.editorial.remaining.threats.mistakesHeading":
-    "Honest mistakes",
-  "publicReview.development.editorial.remaining.threats.mistakes":
-    "An artist can approve the wrong manifest. An operator can choose the wrong address. A governance action can leave out an important value. Ordinary mistakes can cause permanent damage.",
-  "publicReview.development.editorial.remaining.priorities.heading":
-    "Review priorities",
-  "publicReview.development.editorial.remaining.priorities.moneyHeading":
-    "Money and accounting",
-  "publicReview.development.editorial.remaining.priorities.moneyIntro":
-    "Trace every amount owed across sales, auctions, refunds, randomness reserves, split wallets, settlement, and emergency surplus. Prove that:",
-  "publicReview.development.editorial.remaining.priorities.money.received":
-    "Credits never exceed value received.",
-  "publicReview.development.editorial.remaining.priorities.money.promise":
-    "Each unit of value funds only one promise.",
-  "publicReview.development.editorial.remaining.priorities.money.dust":
-    "Rounding leftovers have a clear owner.",
-  "publicReview.development.editorial.remaining.priorities.money.progress":
-    "One recipient that rejects payment cannot block unrelated users.",
-  "publicReview.development.editorial.remaining.priorities.money.emergency":
-    "Emergency withdrawal cannot take money owed to users.",
-  "publicReview.development.editorial.remaining.priorities.money.successor":
-    "A successor change preserves every balance exactly once.",
-  "publicReview.development.editorial.remaining.priorities.money.wallets":
-    "Accounting works with contract wallets and recipients that point back to the protocol.",
-  "publicReview.development.editorial.remaining.priorities.moneyTests":
-    "Invariant tests should combine sales, bids, refunds, withdrawals, pauses, burns, and successor changes.",
-  "publicReview.development.editorial.remaining.priorities.signaturesHeading":
-    "Signed actions",
-  "publicReview.development.editorial.remaining.priorities.signaturesBinding":
-    "Every signature should bind the chain, contract, action, collection, people, money terms, quantity, replay ID, signer version, and deadline.",
-  "publicReview.development.editorial.remaining.priorities.signaturesCompare":
-    "Compare the message shown in the wallet with the Solidity type, recovered signer, replay record, and event. If one value is missing, the user may approve a different action from the one executed.",
-  "publicReview.development.editorial.remaining.priorities.artworkHeading":
-    "Artwork and preservation",
-  "publicReview.development.editorial.remaining.priorities.artworkTest":
-    "Test malformed text, unsafe HTML or JavaScript, large responses, missing files, wrong hashes, unavailable storage, and future browser behavior.",
-  "publicReview.development.editorial.remaining.priorities.artworkReason":
-    "A contract can stay secure while its artwork disappears. Release needs both contract-security proof and preservation proof.",
-  "publicReview.development.editorial.remaining.findings.heading":
-    "Public findings",
-  "publicReview.development.editorial.remaining.findings.policy":
-    "The current disclosure policy allows possible exploitable issues to be reported in the public review Wave while this candidate is in its validated predeployment state.",
-  "publicReview.development.editorial.remaining.findings.guide":
-    "Read the reporting and sensitive-information rules",
-  "publicReview.development.editorial.remaining.findings.response":
-    "Each important response should link to a source commit, test, or recorded decision. A statement alone does not resolve a finding.",
-  "publicReview.development.editorial.remaining.findings.destinationHeading":
-    "Feedback destinations",
-  "publicReview.development.editorial.remaining.findings.staging":
-    "Staging uses Wave 19d4bbf5-86ec-4053-a5f2-bb28d7a2f780, named Stream review (staging).",
-  "publicReview.development.editorial.remaining.findings.production":
-    "The planned production Wave is 06e69198-eea7-40c5-95d3-7c1bf5051aba. Production review routes are still disabled.",
-  "publicReview.development.editorial.remaining.findings.activation":
-    "Before production feedback is enabled, the reviewed environment must bind that exact Wave and a second person must read the mapping back.",
-  "publicReview.development.editorial.remaining.questions.heading":
-    "Questions for reviewers",
-  "publicReview.development.editorial.remaining.questions.progress":
-    "Does this page say that any part is further along than the evidence shows?",
-  "publicReview.development.editorial.remaining.questions.threat":
-    "Which person, asset, dependency, or failure is missing from the threat model?",
-  "publicReview.development.editorial.remaining.questions.static":
-    "Which open static-analysis findings become exploitable when combined?",
-  "publicReview.development.editorial.remaining.questions.bytecode":
-    "Is the chosen Core bytecode margin large enough for a permanent contract?",
-  "publicReview.development.editorial.remaining.questions.governance":
-    "Which governance path still leaves out an important action or value?",
-  "publicReview.development.editorial.remaining.questions.services":
-    "Which real-service tests are required before audit and launch?",
-  "publicReview.development.editorial.remaining.questions.proof":
-    "Which rules need independent economic or formal proof?",
-  "publicReview.development.editorial.remaining.questions.features":
-    "Which planned features are required for launch, and which can wait for a later replacement contract?",
-  "publicReview.development.editorial.proof.heading": "What counts as proof",
-  "publicReview.development.editorial.proof.intro":
-    "Not all proof tells us the same thing. Each step answers a different question.",
-  "publicReview.development.editorial.proof.adrBoundary":
-    "ADR means Architecture Decision Record. It records a design or proposal; it does not prove that the design is built or ready to launch.",
-  "publicReview.development.editorial.proof.proofHeading": "Proof",
-  "publicReview.development.editorial.proof.meaningHeading":
-    "What it tells us",
-  "publicReview.development.editorial.proof.limitHeading":
-    "What it does not prove",
-  "publicReview.development.editorial.proof.code.label":
-    "Code exists in the reviewed version",
-  "publicReview.development.editorial.proof.code.meaning":
-    "The code or generated file exists in the exact version we reviewed.",
-  "publicReview.development.editorial.proof.code.limit":
-    "That it is connected, configured, safe, or meant for launch.",
-  "publicReview.development.editorial.proof.tests.label": "Local tests pass",
-  "publicReview.development.editorial.proof.tests.meaning":
-    "The listed tests passed in the project's test environment.",
-  "publicReview.development.editorial.proof.tests.limit":
-    "That every path is covered or real services work the same way.",
-  "publicReview.development.editorial.proof.setup.label":
-    "Launch setup checked",
-  "publicReview.development.editorial.proof.setup.meaning":
-    "Addresses, code versions, roles, permissions, settings, and connections match one release candidate.",
-  "publicReview.development.editorial.proof.setup.limit":
-    "That marketplaces and other outside services work correctly.",
-  "publicReview.development.editorial.proof.services.label":
-    "Real services tested",
-  "publicReview.development.editorial.proof.services.meaning":
-    "The chosen wallets, marketplaces, public services, storage, and operating processes work outside local tests.",
-  "publicReview.development.editorial.proof.services.limit":
-    "That the full protocol has been reviewed.",
-  "publicReview.development.editorial.proof.audit.label":
-    "Independent audit completed",
-  "publicReview.development.editorial.proof.audit.meaning":
-    "Independent experts reviewed the exact candidate. Findings and fixes were recorded.",
-  "publicReview.development.editorial.proof.audit.limit":
-    "That future changes or daily operations are automatically safe.",
-  "publicReview.development.editorial.proof.current":
-    "Right now, Stream has strong code and local-test proof.",
-  "publicReview.development.editorial.proof.remaining":
-    "Proof for one exact launch setup, real services, deployment, and an independent audit is still incomplete.",
-  "publicReview.development.editorial.technicalHeading": "Technical details",
-  "publicReview.development.editorial.technicalSummary":
-    "All findings on this page refer to one exact version of Stream's code.",
-  "publicReview.development.editorial.commitLabel": "Commit",
-  "publicReview.development.editorial.treeLabel": "Git tree",
-  "publicReview.development.editorial.updateRequired":
-    "If the code changes, this review must be updated.",
-  "publicReview.development.reviewQuestionsHeading":
-    "Where your input would help",
-  "publicReview.development.reviewQuestionsDescription":
-    "Choose the question closest to your experience. Each one explains the issue in plain language and links to the relevant evidence.",
-  "publicReview.development.readQuestion": "Open this question",
-  "publicReview.development.readQuestionLabel": "Open this question: {title}",
+  ...PUBLIC_REVIEW_ARTISTS_AND_ROLES_MESSAGES,
+  ...PUBLIC_REVIEW_ARTWORK_LIFECYCLE_MESSAGES,
+  ...PUBLIC_REVIEW_COMMUNITY_MESSAGES,
+  ...PUBLIC_REVIEW_DEVELOPMENT_MESSAGES,
   "publicReview.evidence.heading": "Evidence labels",
   "publicReview.evidence.summary": "How to read evidence labels",
   "publicReview.evidence.labels.implemented": "Implemented",
